Module lpc55s6x_pac::adc0::cfg

ADC Configuration Register

Structs

PUDLYR
Value of the field

R
Value read from the register

W
Value to write to the register

_HPT_EXDIW
Proxy

_PUDLYW
Proxy

_PWRENW
Proxy

_PWRSELW
Proxy

_REFSELW
Proxy

_TCMDRESW
Proxy

_TPRICTRLW
Proxy

_TRESW
Proxy

Enums

HPT_EXDIR
Possible values of the field HPT_EXDI

HPT_EXDIW
Values that can be written to the field HPT_EXDI

PWRENR
Possible values of the field PWREN

PWRENW
Values that can be written to the field PWREN

PWRSELR
Possible values of the field PWRSEL

PWRSELW
Values that can be written to the field PWRSEL

REFSELR
Possible values of the field REFSEL

REFSELW
Values that can be written to the field REFSEL

TCMDRESR
Possible values of the field TCMDRES

TCMDRESW
Values that can be written to the field TCMDRES

TPRICTRLR
Possible values of the field TPRICTRL

TPRICTRLW
Values that can be written to the field TPRICTRL

TRESR
Possible values of the field TRES

TRESW
Values that can be written to the field TRES