Module lpc55s6x_pac::adc0::swtrig

Software Trigger Register

Structs

R
Value read from the register

W
Value to write to the register

_SWT0W
Proxy

_SWT1W
Proxy

_SWT2W
Proxy

_SWT3W
Proxy

_SWT4W
Proxy

_SWT5W
Proxy

_SWT6W
Proxy

_SWT7W
Proxy

_SWT8W
Proxy

_SWT9W
Proxy

_SWT10W
Proxy

_SWT11W
Proxy

_SWT12W
Proxy

_SWT13W
Proxy

_SWT14W
Proxy

_SWT15W
Proxy

Enums

SWT0R
Possible values of the field SWT0

SWT0W
Values that can be written to the field SWT0

SWT1R
Possible values of the field SWT1

SWT1W
Values that can be written to the field SWT1

SWT2R
Possible values of the field SWT2

SWT2W
Values that can be written to the field SWT2

SWT3R
Possible values of the field SWT3

SWT3W
Values that can be written to the field SWT3

SWT4R
Possible values of the field SWT4

SWT4W
Values that can be written to the field SWT4

SWT5R
Possible values of the field SWT5

SWT5W
Values that can be written to the field SWT5

SWT6R
Possible values of the field SWT6

SWT6W
Values that can be written to the field SWT6

SWT7R
Possible values of the field SWT7

SWT7W
Values that can be written to the field SWT7

SWT8R
Possible values of the field SWT8

SWT8W
Values that can be written to the field SWT8

SWT9R
Possible values of the field SWT9

SWT9W
Values that can be written to the field SWT9

SWT10R
Possible values of the field SWT10

SWT10W
Values that can be written to the field SWT10

SWT11R
Possible values of the field SWT11

SWT11W
Values that can be written to the field SWT11

SWT12R
Possible values of the field SWT12

SWT12W
Values that can be written to the field SWT12

SWT13R
Possible values of the field SWT13

SWT13W
Values that can be written to the field SWT13

SWT14R
Possible values of the field SWT14

SWT14W
Values that can be written to the field SWT14

SWT15R
Possible values of the field SWT15

SWT15W
Values that can be written to the field SWT15