Module lpc55s6x_pac::ahb_secure_ctrl::misc_ctrl_dp_reg

secure control duplicate register

Structs

R
Value read from the register

W
Value to write to the register

_DISABLE_SIMPLE_MASTER_STRICT_MODEW
Proxy

_DISABLE_SMART_MASTER_STRICT_MODEW
Proxy

_DISABLE_VIOLATION_ABORTW
Proxy

_ENABLE_NS_PRIV_CHECKW
Proxy

_ENABLE_SECURE_CHECKINGW
Proxy

_ENABLE_S_PRIV_CHECKW
Proxy

_IDAU_ALL_NSW
Proxy

_WRITE_LOCKW
Proxy

Enums

DISABLE_SIMPLE_MASTER_STRICT_MODER
Possible values of the field DISABLE_SIMPLE_MASTER_STRICT_MODE

DISABLE_SIMPLE_MASTER_STRICT_MODEW
Values that can be written to the field DISABLE_SIMPLE_MASTER_STRICT_MODE

DISABLE_SMART_MASTER_STRICT_MODER
Possible values of the field DISABLE_SMART_MASTER_STRICT_MODE

DISABLE_SMART_MASTER_STRICT_MODEW
Values that can be written to the field DISABLE_SMART_MASTER_STRICT_MODE

DISABLE_VIOLATION_ABORTR
Possible values of the field DISABLE_VIOLATION_ABORT

DISABLE_VIOLATION_ABORTW
Values that can be written to the field DISABLE_VIOLATION_ABORT

ENABLE_NS_PRIV_CHECKR
Possible values of the field ENABLE_NS_PRIV_CHECK

ENABLE_NS_PRIV_CHECKW
Values that can be written to the field ENABLE_NS_PRIV_CHECK

ENABLE_SECURE_CHECKINGR
Possible values of the field ENABLE_SECURE_CHECKING

ENABLE_SECURE_CHECKINGW
Values that can be written to the field ENABLE_SECURE_CHECKING

ENABLE_S_PRIV_CHECKR
Possible values of the field ENABLE_S_PRIV_CHECK

ENABLE_S_PRIV_CHECKW
Values that can be written to the field ENABLE_S_PRIV_CHECK

IDAU_ALL_NSR
Possible values of the field IDAU_ALL_NS

IDAU_ALL_NSW
Values that can be written to the field IDAU_ALL_NS

WRITE_LOCKR
Possible values of the field WRITE_LOCK

WRITE_LOCKW
Values that can be written to the field WRITE_LOCK