Module lpc55s6x_pac::ahb_secure_ctrl::sec_cpu_int_mask0

Secure Interrupt mask for CPU1

Structs

R
Value read from the register

W
Value to write to the register

_ACMP_CAPT0_IRQW
Proxy

_ADC_IRQW
Proxy

_CTIMER0_IRQW
Proxy

_CTIMER1_IRQW
Proxy

_CTIMER3_IRQW
Proxy

_FLEXCOMM0_IRQW
Proxy

_FLEXCOMM1_IRQW
Proxy

_FLEXCOMM2_IRQW
Proxy

_FLEXCOMM3_IRQW
Proxy

_FLEXCOMM4_IRQW
Proxy

_FLEXCOMM5_IRQW
Proxy

_FLEXCOMM6_IRQW
Proxy

_FLEXCOMM7_IRQW
Proxy

_GPIO_GLOBALINT0_IRQW
Proxy

_GPIO_GLOBALINT1_IRQW
Proxy

_GPIO_INT0_IRQ0W
Proxy

_GPIO_INT0_IRQ1W
Proxy

_GPIO_INT0_IRQ2W
Proxy

_GPIO_INT0_IRQ3W
Proxy

_MAILBOX_IRQW
Proxy

_MRT_IRQW
Proxy

_RESERVED0W
Proxy

_RESERVED1W
Proxy

_RESERVED2W
Proxy

_RESERVED3W
Proxy

_RTC_IRQW
Proxy

_SCT_IRQW
Proxy

_SDMA0_IRQW
Proxy

_SYS_IRQW
Proxy

_USB0_IRQW
Proxy

_USB0_NEEDCLKW
Proxy

_UTICK_IRQW
Proxy

Enums

ACMP_CAPT0_IRQR
Possible values of the field ACMP_CAPT0_IRQ

ACMP_CAPT0_IRQW
Values that can be written to the field ACMP_CAPT0_IRQ

ADC_IRQR
Possible values of the field ADC_IRQ

ADC_IRQW
Values that can be written to the field ADC_IRQ

CTIMER0_IRQR
Possible values of the field CTIMER0_IRQ

CTIMER0_IRQW
Values that can be written to the field CTIMER0_IRQ

CTIMER1_IRQR
Possible values of the field CTIMER1_IRQ

CTIMER1_IRQW
Values that can be written to the field CTIMER1_IRQ

CTIMER3_IRQR
Possible values of the field CTIMER3_IRQ

CTIMER3_IRQW
Values that can be written to the field CTIMER3_IRQ

FLEXCOMM0_IRQR
Possible values of the field FLEXCOMM0_IRQ

FLEXCOMM0_IRQW
Values that can be written to the field FLEXCOMM0_IRQ

FLEXCOMM1_IRQR
Possible values of the field FLEXCOMM1_IRQ

FLEXCOMM1_IRQW
Values that can be written to the field FLEXCOMM1_IRQ

FLEXCOMM2_IRQR
Possible values of the field FLEXCOMM2_IRQ

FLEXCOMM2_IRQW
Values that can be written to the field FLEXCOMM2_IRQ

FLEXCOMM3_IRQR
Possible values of the field FLEXCOMM3_IRQ

FLEXCOMM3_IRQW
Values that can be written to the field FLEXCOMM3_IRQ

FLEXCOMM4_IRQR
Possible values of the field FLEXCOMM4_IRQ

FLEXCOMM4_IRQW
Values that can be written to the field FLEXCOMM4_IRQ

FLEXCOMM5_IRQR
Possible values of the field FLEXCOMM5_IRQ

FLEXCOMM5_IRQW
Values that can be written to the field FLEXCOMM5_IRQ

FLEXCOMM6_IRQR
Possible values of the field FLEXCOMM6_IRQ

FLEXCOMM6_IRQW
Values that can be written to the field FLEXCOMM6_IRQ

FLEXCOMM7_IRQR
Possible values of the field FLEXCOMM7_IRQ

FLEXCOMM7_IRQW
Values that can be written to the field FLEXCOMM7_IRQ

GPIO_GLOBALINT0_IRQR
Possible values of the field GPIO_GLOBALINT0_IRQ

GPIO_GLOBALINT0_IRQW
Values that can be written to the field GPIO_GLOBALINT0_IRQ

GPIO_GLOBALINT1_IRQR
Possible values of the field GPIO_GLOBALINT1_IRQ

GPIO_GLOBALINT1_IRQW
Values that can be written to the field GPIO_GLOBALINT1_IRQ

GPIO_INT0_IRQ0R
Possible values of the field GPIO_INT0_IRQ0

GPIO_INT0_IRQ0W
Values that can be written to the field GPIO_INT0_IRQ0

GPIO_INT0_IRQ1R
Possible values of the field GPIO_INT0_IRQ1

GPIO_INT0_IRQ1W
Values that can be written to the field GPIO_INT0_IRQ1

GPIO_INT0_IRQ2R
Possible values of the field GPIO_INT0_IRQ2

GPIO_INT0_IRQ2W
Values that can be written to the field GPIO_INT0_IRQ2

GPIO_INT0_IRQ3R
Possible values of the field GPIO_INT0_IRQ3

GPIO_INT0_IRQ3W
Values that can be written to the field GPIO_INT0_IRQ3

MAILBOX_IRQR
Possible values of the field MAILBOX_IRQ

MAILBOX_IRQW
Values that can be written to the field MAILBOX_IRQ

MRT_IRQR
Possible values of the field MRT_IRQ

MRT_IRQW
Values that can be written to the field MRT_IRQ

RESERVED0R
Possible values of the field RESERVED0

RESERVED0W
Values that can be written to the field RESERVED0

RESERVED1R
Possible values of the field RESERVED1

RESERVED1W
Values that can be written to the field RESERVED1

RESERVED2R
Possible values of the field RESERVED2

RESERVED2W
Values that can be written to the field RESERVED2

RESERVED3R
Possible values of the field RESERVED3

RESERVED3W
Values that can be written to the field RESERVED3

RTC_IRQR
Possible values of the field RTC_IRQ

RTC_IRQW
Values that can be written to the field RTC_IRQ

SCT_IRQR
Possible values of the field SCT_IRQ

SCT_IRQW
Values that can be written to the field SCT_IRQ

SDMA0_IRQR
Possible values of the field SDMA0_IRQ

SDMA0_IRQW
Values that can be written to the field SDMA0_IRQ

SYS_IRQR
Possible values of the field SYS_IRQ

SYS_IRQW
Values that can be written to the field SYS_IRQ

USB0_IRQR
Possible values of the field USB0_IRQ

USB0_IRQW
Values that can be written to the field USB0_IRQ

USB0_NEEDCLKR
Possible values of the field USB0_NEEDCLK

USB0_NEEDCLKW
Values that can be written to the field USB0_NEEDCLK

UTICK_IRQR
Possible values of the field UTICK_IRQ

UTICK_IRQW
Values that can be written to the field UTICK_IRQ