Struct lpc55s6x_pac::ahb_secure_ctrl::sec_cpu_int_mask0::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn sys_irq(&self) -> SYS_IRQR

Bit 0 - Watchdog Timer, Brown Out Detectors and Flash Controller interrupts

pub fn sdma0_irq(&self) -> SDMA0_IRQR

Bit 1 - System DMA 0 (non-secure) interrupt.

pub fn gpio_globalint0_irq(&self) -> GPIO_GLOBALINT0_IRQR

Bit 2 - GPIO Group 0 interrupt.

pub fn gpio_globalint1_irq(&self) -> GPIO_GLOBALINT1_IRQR

Bit 3 - GPIO Group 1 interrupt.

pub fn gpio_int0_irq0(&self) -> GPIO_INT0_IRQ0R

Bit 4 - Pin interrupt 0 or pattern match engine slice 0 interrupt.

pub fn gpio_int0_irq1(&self) -> GPIO_INT0_IRQ1R

Bit 5 - Pin interrupt 1 or pattern match engine slice 1 interrupt.

pub fn gpio_int0_irq2(&self) -> GPIO_INT0_IRQ2R

Bit 6 - Pin interrupt 2 or pattern match engine slice 2 interrupt.

pub fn gpio_int0_irq3(&self) -> GPIO_INT0_IRQ3R

Bit 7 - Pin interrupt 3 or pattern match engine slice 3 interrupt.

pub fn utick_irq(&self) -> UTICK_IRQR

Bit 8 - Micro Tick Timer interrupt.

pub fn mrt_irq(&self) -> MRT_IRQR

Bit 9 - Multi-Rate Timer interrupt.

pub fn ctimer0_irq(&self) -> CTIMER0_IRQR

Bit 10 - Standard counter/timer 0 interrupt.

pub fn ctimer1_irq(&self) -> CTIMER1_IRQR

Bit 11 - Standard counter/timer 1 interrupt.

pub fn sct_irq(&self) -> SCT_IRQR

Bit 12 - SCTimer/PWM interrupt.

pub fn ctimer3_irq(&self) -> CTIMER3_IRQR

Bit 13 - Standard counter/timer 3 interrupt.

pub fn flexcomm0_irq(&self) -> FLEXCOMM0_IRQR

Bit 14 - Flexcomm 0 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm1_irq(&self) -> FLEXCOMM1_IRQR

Bit 15 - Flexcomm 1 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm2_irq(&self) -> FLEXCOMM2_IRQR

Bit 16 - Flexcomm 2 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm3_irq(&self) -> FLEXCOMM3_IRQR

Bit 17 - Flexcomm 3 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm4_irq(&self) -> FLEXCOMM4_IRQR

Bit 18 - Flexcomm 4 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm5_irq(&self) -> FLEXCOMM5_IRQR

Bit 19 - Flexcomm 5 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm6_irq(&self) -> FLEXCOMM6_IRQR

Bit 20 - Flexcomm 6 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm7_irq(&self) -> FLEXCOMM7_IRQR

Bit 21 - Flexcomm 7 interrupt (USART, SPI, I2C, I2S).

pub fn adc_irq(&self) -> ADC_IRQR

Bit 22 - General Purpose ADC interrupt.

pub fn reserved0(&self) -> RESERVED0R

Bit 23 - Reserved. Read value is undefined, only zero should be written.

pub fn acmp_capt0_irq(&self) -> ACMP_CAPT0_IRQR

Bit 24 - Analog Comparator interrupt.

pub fn reserved1(&self) -> RESERVED1R

Bit 25 - Reserved. Read value is undefined, only zero should be written.

pub fn reserved2(&self) -> RESERVED2R

Bit 26 - Reserved. Read value is undefined, only zero should be written.

pub fn usb0_needclk(&self) -> USB0_NEEDCLKR

Bit 27 - USB Full Speed Controller Clock request interrupt.

pub fn usb0_irq(&self) -> USB0_IRQR

Bit 28 - USB High Speed Controller interrupt.

pub fn rtc_irq(&self) -> RTC_IRQR

Bit 29 - RTC_LITE0_ALARM_IRQ, RTC_LITE0_WAKEUP_IRQ

pub fn reserved3(&self) -> RESERVED3R

Bit 30 - Reserved. Read value is undefined, only zero should be written.

pub fn mailbox_irq(&self) -> MAILBOX_IRQR

Bit 31 - Mailbox interrupt.