Module lpc55s6x_pac::flash_cfpa0::dcfg_cc_socu_dflt

With TZ-M, the part can be sold by level 1 customers (secure code developer) to level-2 customers who develops non-secure code only. - In this scenario, or easy of development, Level-I customer releases the part to always allow non-secure debug. - To allow level-2 customers to further seal the part DCFG_CC_SOCU_NS is used. - ROM will use this word to further restrict the debug access.

Structs

INVERSE_VALUER
Value of the field

R
Value read from the register

W
Value to write to the register

_DBGENW
Proxy

_FA_CMD_ENW
Proxy

_INVERSE_VALUEW
Proxy

_ISP_CMD_ENW
Proxy

_MCM33_DBGENW
Proxy

_MCM33_NIDENW
Proxy

_ME_CMD_ENW
Proxy

_NIDENW
Proxy

_SPIDENW
Proxy

_SPNIDENW
Proxy

_TAPENW
Proxy

Enums

DBGENR
Possible values of the field DBGEN

DBGENW
Values that can be written to the field DBGEN

FA_CMD_ENR
Possible values of the field FA_CMD_EN

FA_CMD_ENW
Values that can be written to the field FA_CMD_EN

ISP_CMD_ENR
Possible values of the field ISP_CMD_EN

ISP_CMD_ENW
Values that can be written to the field ISP_CMD_EN

MCM33_DBGENR
Possible values of the field MCM33_DBGEN

MCM33_NIDENR
Possible values of the field MCM33_NIDEN

MCM33_DBGENW
Values that can be written to the field MCM33_DBGEN

MCM33_NIDENW
Values that can be written to the field MCM33_NIDEN

ME_CMD_ENR
Possible values of the field ME_CMD_EN

ME_CMD_ENW
Values that can be written to the field ME_CMD_EN

NIDENR
Possible values of the field NIDEN

NIDENW
Values that can be written to the field NIDEN

SPIDENR
Possible values of the field SPIDEN

SPIDENW
Values that can be written to the field SPIDEN

SPNIDENR
Possible values of the field SPNIDEN

SPNIDENW
Values that can be written to the field SPNIDEN

TAPENR
Possible values of the field TAPEN

TAPENW
Values that can be written to the field TAPEN