Struct lpc55s6x_pac::hashcrypt::RegisterBlock

#[repr(C)]

pub struct RegisterBlock {
    pub ctrl: CTRL,
    pub status: STATUS,
    pub intenset: INTENSET,
    pub intenclr: INTENCLR,
    pub memctrl: MEMCTRL,
    pub memaddr: MEMADDR,
    pub indata: INDATA,
    pub alias: [ALIAS; 7],
    pub outdata0: [OUTDATA0; 8],
    pub outdata1: [OUTDATA1; 8],
    pub cryptcfg: CRYPTCFG,
    pub config: CONFIG,
    pub lock: LOCK,
    pub mask: [MASK; 4],
    // some fields omitted
}

Register block

Fields
ctrl: CTRL

0x00 - Is control register to enable and operate Hash and Crypto

status: STATUS

0x04 - Indicates status of Hash peripheral.

intenset: INTENSET

0x08 - Write 1 to enable interrupts; reads back with which are set.

intenclr: INTENCLR

0x0c - Write 1 to clear interrupts.

memctrl: MEMCTRL

0x10 - Setup Master to access memory (if available)

memaddr: MEMADDR

0x14 - Address to start memory access from (if available).

indata: INDATA

0x20 - Input of 16 words at a time to load up buffer.

alias: [ALIAS; 7]

0x24 - no description available

outdata0: [OUTDATA0; 8]

0x40 - no description available

outdata1: [OUTDATA1; 8]

0x60 - no description available

cryptcfg: CRYPTCFG

0x80 - Crypto settings for AES and Salsa and ChaCha

config: CONFIG

0x84 - Returns the configuration of this block in this chip - indicates what services are available.

lock: LOCK

0x8c - Lock register allows locking to the current security level or unlocking by the lock holding level.

mask: [MASK; 4]

0x90 - no description available