Module lpc55s6x_pac::i2c0::cfg

Configuration for shared functions.

Structs

R
Value read from the register

W
Value to write to the register

_HSCAPABLEW
Proxy

_MONCLKSTRW
Proxy

_MONENW
Proxy

_MSTENW
Proxy

_SLVENW
Proxy

_TIMEOUTENW
Proxy

Enums

HSCAPABLER
Possible values of the field HSCAPABLE

HSCAPABLEW
Values that can be written to the field HSCAPABLE

MONCLKSTRR
Possible values of the field MONCLKSTR

MONCLKSTRW
Values that can be written to the field MONCLKSTR

MONENR
Possible values of the field MONEN

MONENW
Values that can be written to the field MONEN

MSTENR
Possible values of the field MSTEN

MSTENW
Values that can be written to the field MSTEN

SLVENR
Possible values of the field SLVEN

SLVENW
Values that can be written to the field SLVEN

TIMEOUTENR
Possible values of the field TIMEOUTEN

TIMEOUTENW
Values that can be written to the field TIMEOUTEN