Module lpc55s6x_pac::i2c0::intenset

Interrupt Enable Set and read register.

Structs

R
Value read from the register

W
Value to write to the register

_EVENTTIMEOUTENW
Proxy

_MONIDLEENW
Proxy

_MONOVENW
Proxy

_MONRDYENW
Proxy

_MSTARBLOSSENW
Proxy

_MSTPENDINGENW
Proxy

_MSTSTSTPERRENW
Proxy

_SCLTIMEOUTENW
Proxy

_SLVDESELENW
Proxy

_SLVNOTSTRENW
Proxy

_SLVPENDINGENW
Proxy

Enums

EVENTTIMEOUTENR
Possible values of the field EVENTTIMEOUTEN

EVENTTIMEOUTENW
Values that can be written to the field EVENTTIMEOUTEN

MONIDLEENR
Possible values of the field MONIDLEEN

MONIDLEENW
Values that can be written to the field MONIDLEEN

MONOVENR
Possible values of the field MONOVEN

MONOVENW
Values that can be written to the field MONOVEN

MONRDYENR
Possible values of the field MONRDYEN

MONRDYENW
Values that can be written to the field MONRDYEN

MSTARBLOSSENR
Possible values of the field MSTARBLOSSEN

MSTARBLOSSENW
Values that can be written to the field MSTARBLOSSEN

MSTPENDINGENR
Possible values of the field MSTPENDINGEN

MSTPENDINGENW
Values that can be written to the field MSTPENDINGEN

MSTSTSTPERRENR
Possible values of the field MSTSTSTPERREN

MSTSTSTPERRENW
Values that can be written to the field MSTSTSTPERREN

SCLTIMEOUTENR
Possible values of the field SCLTIMEOUTEN

SCLTIMEOUTENW
Values that can be written to the field SCLTIMEOUTEN

SLVDESELENR
Possible values of the field SLVDESELEN

SLVDESELENW
Values that can be written to the field SLVDESELEN

SLVNOTSTRENR
Possible values of the field SLVNOTSTREN

SLVNOTSTRENW
Values that can be written to the field SLVNOTSTREN

SLVPENDINGENR
Possible values of the field SLVPENDINGEN

SLVPENDINGENW
Values that can be written to the field SLVPENDINGEN