Module lpc55s6x_pac::i2s0::cfg1

Configuration register 1 for the primary channel pair.

Structs

DATALENR
Value of the field

R
Value read from the register

W
Value to write to the register

_DATALENW
Proxy

_DATAPAUSEW
Proxy

_LEFTJUSTW
Proxy

_MAINENABLEW
Proxy

_MODEW
Proxy

_MSTSLVCFGW
Proxy

_ONECHANNELW
Proxy

_PAIRCOUNTW
Proxy

_PDMDATAW
Proxy

_RIGHTLOWW
Proxy

_SCK_POLW
Proxy

_WS_POLW
Proxy

Enums

DATAPAUSER
Possible values of the field DATAPAUSE

DATAPAUSEW
Values that can be written to the field DATAPAUSE

LEFTJUSTR
Possible values of the field LEFTJUST

LEFTJUSTW
Values that can be written to the field LEFTJUST

MAINENABLER
Possible values of the field MAINENABLE

MAINENABLEW
Values that can be written to the field MAINENABLE

MODER
Possible values of the field MODE

MODEW
Values that can be written to the field MODE

MSTSLVCFGR
Possible values of the field MSTSLVCFG

MSTSLVCFGW
Values that can be written to the field MSTSLVCFG

ONECHANNELR
Possible values of the field ONECHANNEL

ONECHANNELW
Values that can be written to the field ONECHANNEL

PAIRCOUNTR
Possible values of the field PAIRCOUNT

PAIRCOUNTW
Values that can be written to the field PAIRCOUNT

PDMDATAR
Possible values of the field PDMDATA

PDMDATAW
Values that can be written to the field PDMDATA

RIGHTLOWR
Possible values of the field RIGHTLOW

RIGHTLOWW
Values that can be written to the field RIGHTLOW

SCK_POLR
Possible values of the field SCK_POL

SCK_POLW
Values that can be written to the field SCK_POL

WS_POLR
Possible values of the field WS_POL

WS_POLW
Values that can be written to the field WS_POL