Enum lpc55s6x_pac::inputmux::timer4captsel::CAPTSELR

pub enum CAPTSELR {
    VAL0,
    VAL1,
    VAL2,
    VAL3,
    VAL4,
    VAL5,
    VAL6,
    VAL7,
    VAL8,
    VAL9,
    VAL10,
    VAL11,
    VAL12,
    VAL13,
    VAL14,
    VAL15,
    VAL16,
    VAL17,
    VAL18,
    VAL19,
    VAL20,
    VAL21,
    VAL22,
    VAL23,
    VAL24,
    VAL25,
    _Reserved(u8),
}

Possible values of the field CAPTSEL

Variants
VAL0

CT_INP0 function selected from IOCON register

VAL1

CT_INP1 function selected from IOCON register

VAL2

CT_INP2 function selected from IOCON register

VAL3

CT_INP3 function selected from IOCON register

VAL4

CT_INP4 function selected from IOCON register

VAL5

CT_INP5 function selected from IOCON register

VAL6

CT_INP6 function selected from IOCON register

VAL7

CT_INP7 function selected from IOCON register

VAL8

CT_INP8 function selected from IOCON register

VAL9

CT_INP9 function selected from IOCON register

VAL10

CT_INP10 function selected from IOCON register

VAL11

CT_INP11 function selected from IOCON register

VAL12

CT_INP12 function selected from IOCON register

VAL13

CT_INP13 function selected from IOCON register

VAL14

CT_INP14 function selected from IOCON register

VAL15

CT_INP15 function selected from IOCON register

VAL16

CT_INP16 function selected from IOCON register

VAL17

CT_INP17 function selected from IOCON register

VAL18

CT_INP18 function selected from IOCON register

VAL19

CT_INP19 function selected from IOCON register

VAL20

USB0_FRAME_TOGGLE

VAL21

USB1_FRAME_TOGGLE

VAL22

COMP_OUTPUT output from analog comparator

VAL23

I2S_SHARED_WS[0] output from I2S pin sharing

VAL24

I2S_SHARED_WS[1] output from I2S pin sharing

VAL25

None

_Reserved(u8)

Reserved

Methods

impl CAPTSELR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_val0(&self) -> bool

Checks if the value of the field is VAL0

pub fn is_val1(&self) -> bool

Checks if the value of the field is VAL1

pub fn is_val2(&self) -> bool

Checks if the value of the field is VAL2

pub fn is_val3(&self) -> bool

Checks if the value of the field is VAL3

pub fn is_val4(&self) -> bool

Checks if the value of the field is VAL4

pub fn is_val5(&self) -> bool

Checks if the value of the field is VAL5

pub fn is_val6(&self) -> bool

Checks if the value of the field is VAL6

pub fn is_val7(&self) -> bool

Checks if the value of the field is VAL7

pub fn is_val8(&self) -> bool

Checks if the value of the field is VAL8

pub fn is_val9(&self) -> bool

Checks if the value of the field is VAL9

pub fn is_val10(&self) -> bool

Checks if the value of the field is VAL10

pub fn is_val11(&self) -> bool

Checks if the value of the field is VAL11

pub fn is_val12(&self) -> bool

Checks if the value of the field is VAL12

pub fn is_val13(&self) -> bool

Checks if the value of the field is VAL13

pub fn is_val14(&self) -> bool

Checks if the value of the field is VAL14

pub fn is_val15(&self) -> bool

Checks if the value of the field is VAL15

pub fn is_val16(&self) -> bool

Checks if the value of the field is VAL16

pub fn is_val17(&self) -> bool

Checks if the value of the field is VAL17

pub fn is_val18(&self) -> bool

Checks if the value of the field is VAL18

pub fn is_val19(&self) -> bool

Checks if the value of the field is VAL19

pub fn is_val20(&self) -> bool

Checks if the value of the field is VAL20

pub fn is_val21(&self) -> bool

Checks if the value of the field is VAL21

pub fn is_val22(&self) -> bool

Checks if the value of the field is VAL22

pub fn is_val23(&self) -> bool

Checks if the value of the field is VAL23

pub fn is_val24(&self) -> bool

Checks if the value of the field is VAL24

pub fn is_val25(&self) -> bool

Checks if the value of the field is VAL25