Module lpc55s6x_pac::pint::pmcfg

Pattern match interrupt bit slice configuration register

Structs

R
Value read from the register

W
Value to write to the register

_CFG0W
Proxy

_CFG1W
Proxy

_CFG2W
Proxy

_CFG3W
Proxy

_CFG4W
Proxy

_CFG5W
Proxy

_CFG6W
Proxy

_CFG7W
Proxy

_PROD_ENDPTS0W
Proxy

_PROD_ENDPTS1W
Proxy

_PROD_ENDPTS2W
Proxy

_PROD_ENDPTS3W
Proxy

_PROD_ENDPTS4W
Proxy

_PROD_ENDPTS5W
Proxy

_PROD_ENDPTS6W
Proxy

Enums

CFG0R
Possible values of the field CFG0

CFG0W
Values that can be written to the field CFG0

CFG1R
Possible values of the field CFG1

CFG1W
Values that can be written to the field CFG1

CFG2R
Possible values of the field CFG2

CFG2W
Values that can be written to the field CFG2

CFG3R
Possible values of the field CFG3

CFG3W
Values that can be written to the field CFG3

CFG4R
Possible values of the field CFG4

CFG4W
Values that can be written to the field CFG4

CFG5R
Possible values of the field CFG5

CFG5W
Values that can be written to the field CFG5

CFG6R
Possible values of the field CFG6

CFG6W
Values that can be written to the field CFG6

CFG7R
Possible values of the field CFG7

CFG7W
Values that can be written to the field CFG7

PROD_ENDPTS0R
Possible values of the field PROD_ENDPTS0

PROD_ENDPTS0W
Values that can be written to the field PROD_ENDPTS0

PROD_ENDPTS1R
Possible values of the field PROD_ENDPTS1

PROD_ENDPTS1W
Values that can be written to the field PROD_ENDPTS1

PROD_ENDPTS2R
Possible values of the field PROD_ENDPTS2

PROD_ENDPTS2W
Values that can be written to the field PROD_ENDPTS2

PROD_ENDPTS3R
Possible values of the field PROD_ENDPTS3

PROD_ENDPTS3W
Values that can be written to the field PROD_ENDPTS3

PROD_ENDPTS4R
Possible values of the field PROD_ENDPTS4

PROD_ENDPTS4W
Values that can be written to the field PROD_ENDPTS4

PROD_ENDPTS5R
Possible values of the field PROD_ENDPTS5

PROD_ENDPTS5W
Values that can be written to the field PROD_ENDPTS5

PROD_ENDPTS6R
Possible values of the field PROD_ENDPTS6

PROD_ENDPTS6W
Values that can be written to the field PROD_ENDPTS6