Enum lpc55s6x_pac::plu::lut::lut_inp::LUT_INPR

pub enum LUT_INPR {
    PLU_INPUTS0,
    PLU_INPUTS1,
    PLU_INPUTS2,
    PLU_INPUTS3,
    PLU_INPUTS4,
    PLU_INPUTS5,
    LUT_OUTPUTS0,
    LUT_OUTPUTS1,
    LUT_OUTPUTS2,
    LUT_OUTPUTS3,
    LUT_OUTPUTS4,
    LUT_OUTPUTS5,
    LUT_OUTPUTS6,
    LUT_OUTPUTS7,
    LUT_OUTPUTS8,
    LUT_OUTPUTS9,
    LUT_OUTPUTS10,
    LUT_OUTPUTS11,
    LUT_OUTPUTS12,
    LUT_OUTPUTS13,
    LUT_OUTPUTS14,
    LUT_OUTPUTS15,
    LUT_OUTPUTS16,
    LUT_OUTPUTS17,
    LUT_OUTPUTS18,
    LUT_OUTPUTS19,
    LUT_OUTPUTS20,
    LUT_OUTPUTS21,
    LUT_OUTPUTS22,
    LUT_OUTPUTS23,
    LUT_OUTPUTS24,
    LUT_OUTPUTS25,
    STATE0,
    STATE1,
    STATE2,
    STATE3,
    _Reserved(u8),
}

Possible values of the field LUT_INP

Variants
PLU_INPUTS0

The PLU primary inputs 0.

PLU_INPUTS1

The PLU primary inputs 1.

PLU_INPUTS2

The PLU primary inputs 2.

PLU_INPUTS3

The PLU primary inputs 3.

PLU_INPUTS4

The PLU primary inputs 4.

PLU_INPUTS5

The PLU primary inputs 5.

LUT_OUTPUTS0

The output of LUT0.

LUT_OUTPUTS1

The output of LUT1.

LUT_OUTPUTS2

The output of LUT2.

LUT_OUTPUTS3

The output of LUT3.

LUT_OUTPUTS4

The output of LUT4.

LUT_OUTPUTS5

The output of LUT5.

LUT_OUTPUTS6

The output of LUT6.

LUT_OUTPUTS7

The output of LUT7.

LUT_OUTPUTS8

The output of LUT8.

LUT_OUTPUTS9

The output of LUT9.

LUT_OUTPUTS10

The output of LUT10.

LUT_OUTPUTS11

The output of LUT11.

LUT_OUTPUTS12

The output of LUT12.

LUT_OUTPUTS13

The output of LUT13.

LUT_OUTPUTS14

The output of LUT14.

LUT_OUTPUTS15

The output of LUT15.

LUT_OUTPUTS16

The output of LUT16.

LUT_OUTPUTS17

The output of LUT17.

LUT_OUTPUTS18

The output of LUT18.

LUT_OUTPUTS19

The output of LUT19.

LUT_OUTPUTS20

The output of LUT20.

LUT_OUTPUTS21

The output of LUT21.

LUT_OUTPUTS22

The output of LUT22.

LUT_OUTPUTS23

The output of LUT23.

LUT_OUTPUTS24

The output of LUT24.

LUT_OUTPUTS25

The output of LUT25.

STATE0

state(0).

STATE1

state(1).

STATE2

state(2).

STATE3

state(3).

_Reserved(u8)

Reserved

Methods

impl LUT_INPR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_plu_inputs0(&self) -> bool

Checks if the value of the field is PLU_INPUTS0

pub fn is_plu_inputs1(&self) -> bool

Checks if the value of the field is PLU_INPUTS1

pub fn is_plu_inputs2(&self) -> bool

Checks if the value of the field is PLU_INPUTS2

pub fn is_plu_inputs3(&self) -> bool

Checks if the value of the field is PLU_INPUTS3

pub fn is_plu_inputs4(&self) -> bool

Checks if the value of the field is PLU_INPUTS4

pub fn is_plu_inputs5(&self) -> bool

Checks if the value of the field is PLU_INPUTS5

pub fn is_lut_outputs0(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS0

pub fn is_lut_outputs1(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS1

pub fn is_lut_outputs2(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS2

pub fn is_lut_outputs3(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS3

pub fn is_lut_outputs4(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS4

pub fn is_lut_outputs5(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS5

pub fn is_lut_outputs6(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS6

pub fn is_lut_outputs7(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS7

pub fn is_lut_outputs8(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS8

pub fn is_lut_outputs9(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS9

pub fn is_lut_outputs10(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS10

pub fn is_lut_outputs11(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS11

pub fn is_lut_outputs12(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS12

pub fn is_lut_outputs13(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS13

pub fn is_lut_outputs14(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS14

pub fn is_lut_outputs15(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS15

pub fn is_lut_outputs16(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS16

pub fn is_lut_outputs17(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS17

pub fn is_lut_outputs18(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS18

pub fn is_lut_outputs19(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS19

pub fn is_lut_outputs20(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS20

pub fn is_lut_outputs21(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS21

pub fn is_lut_outputs22(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS22

pub fn is_lut_outputs23(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS23

pub fn is_lut_outputs24(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS24

pub fn is_lut_outputs25(&self) -> bool

Checks if the value of the field is LUT_OUTPUTS25

pub fn is_state0(&self) -> bool

Checks if the value of the field is STATE0

pub fn is_state1(&self) -> bool

Checks if the value of the field is STATE1

pub fn is_state2(&self) -> bool

Checks if the value of the field is STATE2

pub fn is_state3(&self) -> bool

Checks if the value of the field is STATE3