Module lpc55s6x_pac::pmc

PMC

Modules

aoreg1
General purpose always on domain data storage [Reset by: PoR, Brown Out Detectors Reset]

bodcore
Digital Core logic Brown Out Dectector control register [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

bodvbat
VBAT Brown Out Dectector (BoD) control register [Reset by: PoR, Pin Reset, Software Reset]

comp
Analog Comparator control register [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

fro1m
1 MHz Free Running Oscillator control register [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

fro32k
32 KHz Free Running Oscillator (FRO) control register [Reset by: PoR, Brown Out Detectors Reset]

ostimer
OS Timer control register [Reset by: PoR, Brown Out Detectors Reset]

pdruncfg0
Controls the power to various analog blocks [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

pdruncfgclr0
Controls the power to various analog blocks [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

pdruncfgset0
Controls the power to various analog blocks [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

pdsleepcfg0
Controls the power to various modules during Low Power modes - DEEP SLEEP, POWER DOWN and DEEP POWER DOWN [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Software Reset]

resetcause
Reset Cause register [Reset by: PoR]

resetctrl
Reset Control [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

rtcosc32k
RTC 1 KHZ and 1 Hz clocks source control register [Reset by: PoR, Brown Out Detectors Reset]

statusclk
FRO and XTAL status register [Reset by: PoR, Brown Out Detectors Reset]

wakeiocause
Allows to identify the Wake-up I/O source from Deep Power Down mode

xtal32k
32 KHz Crystal oscillator (XTAL) control register [Reset by: PoR, Brown Out Detectors Reset]

Structs

AOREG1
General purpose always on domain data storage [Reset by: PoR, Brown Out Detectors Reset]

BODCORE
Digital Core logic Brown Out Dectector control register [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

BODVBAT
VBAT Brown Out Dectector (BoD) control register [Reset by: PoR, Pin Reset, Software Reset]

COMP
Analog Comparator control register [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

FRO1M
1 MHz Free Running Oscillator control register [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

FRO32K
32 KHz Free Running Oscillator (FRO) control register [Reset by: PoR, Brown Out Detectors Reset]

OSTIMER
OS Timer control register [Reset by: PoR, Brown Out Detectors Reset]

PDRUNCFG0
Controls the power to various analog blocks [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

PDRUNCFGCLR0
Controls the power to various analog blocks [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

PDRUNCFGSET0
Controls the power to various analog blocks [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

PDSLEEPCFG0
Controls the power to various modules during Low Power modes - DEEP SLEEP, POWER DOWN and DEEP POWER DOWN [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Software Reset]

RESETCAUSE
Reset Cause register [Reset by: PoR]

RESETCTRL
Reset Control [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

RTCOSC32K
RTC 1 KHZ and 1 Hz clocks source control register [Reset by: PoR, Brown Out Detectors Reset]

RegisterBlock
Register block

STATUSCLK
FRO and XTAL status register [Reset by: PoR, Brown Out Detectors Reset]

WAKEIOCAUSE
Allows to identify the Wake-up I/O source from Deep Power Down mode

XTAL32K
32 KHz Crystal oscillator (XTAL) control register [Reset by: PoR, Brown Out Detectors Reset]