Enum lpc55s6x_pac::pmc::fro1m::DIVSELR

pub enum DIVSELR {
    DIV_2,
    DIV_4,
    DIV_6,
    DIV_8,
    DIV_10,
    DIV_12,
    DIV_14,
    DIV_16,
    DIV_18,
    DIV_20,
    DIV_22,
    DIV_24,
    DIV_26,
    DIV_28,
    DIV_30,
    DIV_32,
    DIV_34,
    DIV_36,
    DIV_38,
    DIV_40,
    DIV_42,
    DIV_44,
    DIV_46,
    DIV_48,
    DIV_50,
    DIV_52,
    DIV_54,
    DIV_56,
    DIV_58,
    DIV_60,
    DIV_62,
    DIV_1,
}

Possible values of the field DIVSEL

Variants
DIV_2

2.0.

DIV_4

4.0.

DIV_6

6.0.

DIV_8

8.0.

DIV_10

10.0.

DIV_12

12.0.

DIV_14

14.0.

DIV_16

16.0.

DIV_18

18.0.

DIV_20

20.0.

DIV_22

22.0.

DIV_24

24.0.

DIV_26

26.0.

DIV_28

28.0.

DIV_30

30.0.

DIV_32

32.0.

DIV_34

34.0.

DIV_36

36.0.

DIV_38

38.0.

DIV_40

40.0.

DIV_42

42.0.

DIV_44

44.0.

DIV_46

46.0.

DIV_48

48.0.

DIV_50

50.0.

DIV_52

52.0.

DIV_54

54.0.

DIV_56

56.0.

DIV_58

58.0.

DIV_60

60.0.

DIV_62

62.0.

DIV_1

1.0.

Methods

impl DIVSELR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_div_2(&self) -> bool

Checks if the value of the field is DIV_2

pub fn is_div_4(&self) -> bool

Checks if the value of the field is DIV_4

pub fn is_div_6(&self) -> bool

Checks if the value of the field is DIV_6

pub fn is_div_8(&self) -> bool

Checks if the value of the field is DIV_8

pub fn is_div_10(&self) -> bool

Checks if the value of the field is DIV_10

pub fn is_div_12(&self) -> bool

Checks if the value of the field is DIV_12

pub fn is_div_14(&self) -> bool

Checks if the value of the field is DIV_14

pub fn is_div_16(&self) -> bool

Checks if the value of the field is DIV_16

pub fn is_div_18(&self) -> bool

Checks if the value of the field is DIV_18

pub fn is_div_20(&self) -> bool

Checks if the value of the field is DIV_20

pub fn is_div_22(&self) -> bool

Checks if the value of the field is DIV_22

pub fn is_div_24(&self) -> bool

Checks if the value of the field is DIV_24

pub fn is_div_26(&self) -> bool

Checks if the value of the field is DIV_26

pub fn is_div_28(&self) -> bool

Checks if the value of the field is DIV_28

pub fn is_div_30(&self) -> bool

Checks if the value of the field is DIV_30

pub fn is_div_32(&self) -> bool

Checks if the value of the field is DIV_32

pub fn is_div_34(&self) -> bool

Checks if the value of the field is DIV_34

pub fn is_div_36(&self) -> bool

Checks if the value of the field is DIV_36

pub fn is_div_38(&self) -> bool

Checks if the value of the field is DIV_38

pub fn is_div_40(&self) -> bool

Checks if the value of the field is DIV_40

pub fn is_div_42(&self) -> bool

Checks if the value of the field is DIV_42

pub fn is_div_44(&self) -> bool

Checks if the value of the field is DIV_44

pub fn is_div_46(&self) -> bool

Checks if the value of the field is DIV_46

pub fn is_div_48(&self) -> bool

Checks if the value of the field is DIV_48

pub fn is_div_50(&self) -> bool

Checks if the value of the field is DIV_50

pub fn is_div_52(&self) -> bool

Checks if the value of the field is DIV_52

pub fn is_div_54(&self) -> bool

Checks if the value of the field is DIV_54

pub fn is_div_56(&self) -> bool

Checks if the value of the field is DIV_56

pub fn is_div_58(&self) -> bool

Checks if the value of the field is DIV_58

pub fn is_div_60(&self) -> bool

Checks if the value of the field is DIV_60

pub fn is_div_62(&self) -> bool

Checks if the value of the field is DIV_62

pub fn is_div_1(&self) -> bool

Checks if the value of the field is DIV_1