Module lpc55s6x_pac::pmc::pdruncfg0

Controls the power to various analog blocks [Reset by: PoR, Pin Reset, Brown Out Detectors Reset, Deep Power Down Reset, Software Reset]

Structs

R
Value read from the register

W
Value to write to the register

_PDEN_AUXBIASW
Proxy

_PDEN_BIASW
Proxy

_PDEN_BODCOREW
Proxy

_PDEN_BODVBATW
Proxy

_PDEN_COMPW
Proxy

_PDEN_DCDCW
Proxy

_PDEN_FRO32KW
Proxy

_PDEN_FRO192MW
Proxy

_PDEN_GPADCW
Proxy

_PDEN_LDODEEPSLEEPW
Proxy

_PDEN_LDOFLASHNVW
Proxy

_PDEN_LDOMEMW
Proxy

_PDEN_LDOUSBHSW
Proxy

_PDEN_LDOXO32MW
Proxy

_PDEN_PLL0W
Proxy

_PDEN_PLL0_SSCGW
Proxy

_PDEN_PLL1W
Proxy

_PDEN_RNGW
Proxy

_PDEN_TEMPSENSW
Proxy

_PDEN_USBFSPHYW
Proxy

_PDEN_USBHSPHYW
Proxy

_PDEN_XTAL32KW
Proxy

_PDEN_XTAL32MW
Proxy

Enums

PDEN_AUXBIASR
Possible values of the field PDEN_AUXBIAS

PDEN_AUXBIASW
Values that can be written to the field PDEN_AUXBIAS

PDEN_BIASR
Possible values of the field PDEN_BIAS

PDEN_BIASW
Values that can be written to the field PDEN_BIAS

PDEN_BODCORER
Possible values of the field PDEN_BODCORE

PDEN_BODCOREW
Values that can be written to the field PDEN_BODCORE

PDEN_BODVBATR
Possible values of the field PDEN_BODVBAT

PDEN_BODVBATW
Values that can be written to the field PDEN_BODVBAT

PDEN_COMPR
Possible values of the field PDEN_COMP

PDEN_COMPW
Values that can be written to the field PDEN_COMP

PDEN_DCDCR
Possible values of the field PDEN_DCDC

PDEN_DCDCW
Values that can be written to the field PDEN_DCDC

PDEN_FRO32KR
Possible values of the field PDEN_FRO32K

PDEN_FRO32KW
Values that can be written to the field PDEN_FRO32K

PDEN_FRO192MR
Possible values of the field PDEN_FRO192M

PDEN_FRO192MW
Values that can be written to the field PDEN_FRO192M

PDEN_GPADCR
Possible values of the field PDEN_GPADC

PDEN_GPADCW
Values that can be written to the field PDEN_GPADC

PDEN_LDODEEPSLEEPR
Possible values of the field PDEN_LDODEEPSLEEP

PDEN_LDODEEPSLEEPW
Values that can be written to the field PDEN_LDODEEPSLEEP

PDEN_LDOFLASHNVR
Possible values of the field PDEN_LDOFLASHNV

PDEN_LDOFLASHNVW
Values that can be written to the field PDEN_LDOFLASHNV

PDEN_LDOMEMR
Possible values of the field PDEN_LDOMEM

PDEN_LDOMEMW
Values that can be written to the field PDEN_LDOMEM

PDEN_LDOUSBHSR
Possible values of the field PDEN_LDOUSBHS

PDEN_LDOUSBHSW
Values that can be written to the field PDEN_LDOUSBHS

PDEN_LDOXO32MR
Possible values of the field PDEN_LDOXO32M

PDEN_LDOXO32MW
Values that can be written to the field PDEN_LDOXO32M

PDEN_PLL0R
Possible values of the field PDEN_PLL0

PDEN_PLL0W
Values that can be written to the field PDEN_PLL0

PDEN_PLL0_SSCGR
Possible values of the field PDEN_PLL0_SSCG

PDEN_PLL0_SSCGW
Values that can be written to the field PDEN_PLL0_SSCG

PDEN_PLL1R
Possible values of the field PDEN_PLL1

PDEN_PLL1W
Values that can be written to the field PDEN_PLL1

PDEN_RNGR
Possible values of the field PDEN_RNG

PDEN_RNGW
Values that can be written to the field PDEN_RNG

PDEN_TEMPSENSR
Possible values of the field PDEN_TEMPSENS

PDEN_TEMPSENSW
Values that can be written to the field PDEN_TEMPSENS

PDEN_USBFSPHYR
Possible values of the field PDEN_USBFSPHY

PDEN_USBFSPHYW
Values that can be written to the field PDEN_USBFSPHY

PDEN_USBHSPHYR
Possible values of the field PDEN_USBHSPHY

PDEN_USBHSPHYW
Values that can be written to the field PDEN_USBHSPHY

PDEN_XTAL32KR
Possible values of the field PDEN_XTAL32K

PDEN_XTAL32MR
Possible values of the field PDEN_XTAL32M

PDEN_XTAL32KW
Values that can be written to the field PDEN_XTAL32K

PDEN_XTAL32MW
Values that can be written to the field PDEN_XTAL32M