Module lpc55s6x_pac::pmc::xtal32k

32 KHz Crystal oscillator (XTAL) control register [Reset by: PoR, Brown Out Detectors Reset]

Structs

AMPLR
Value of the field

CAPBANKINR
Value of the field

CAPBANKOUTR
Value of the field

CAPTESTENABLER
Value of the field

CAPTESTSTARTR
Value of the field

IBIASR
Value of the field

IREFR
Value of the field

R
Value read from the register

TESTR
Value of the field

W
Value to write to the register

_AMPLW
Proxy

_CAPBANKINW
Proxy

_CAPBANKOUTW
Proxy

_CAPTESTENABLEW
Proxy

_CAPTESTOSCINSELW
Proxy

_CAPTESTSTARTSRCSELW
Proxy

_CAPTESTSTARTW
Proxy

_IBIASW
Proxy

_IREFW
Proxy

_TESTW
Proxy

Enums

CAPTESTOSCINSELR
Possible values of the field CAPTESTOSCINSEL

CAPTESTOSCINSELW
Values that can be written to the field CAPTESTOSCINSEL

CAPTESTSTARTSRCSELR
Possible values of the field CAPTESTSTARTSRCSEL

CAPTESTSTARTSRCSELW
Values that can be written to the field CAPTESTSTARTSRCSEL