Module lpc55s6x_pac::sau::sfsr

Secure Fault Status Register

Structs

R
Value read from the register

W
Value to write to the register

_AUVIOLW
Proxy

_INVEPW
Proxy

_INVERW
Proxy

_INVISW
Proxy

_INVTRANW
Proxy

_LSERRW
Proxy

_LSPERRW
Proxy

_SFARVALIDW
Proxy

Enums

AUVIOLR
Possible values of the field AUVIOL

AUVIOLW
Values that can be written to the field AUVIOL

INVEPR
Possible values of the field INVEP

INVEPW
Values that can be written to the field INVEP

INVERR
Possible values of the field INVER

INVERW
Values that can be written to the field INVER

INVISR
Possible values of the field INVIS

INVISW
Values that can be written to the field INVIS

INVTRANR
Possible values of the field INVTRAN

INVTRANW
Values that can be written to the field INVTRAN

LSERRR
Possible values of the field LSERR

LSERRW
Values that can be written to the field LSERR

LSPERRR
Possible values of the field LSPERR

LSPERRW
Values that can be written to the field LSPERR

SFARVALIDR
Possible values of the field SFARVALID

SFARVALIDW
Values that can be written to the field SFARVALID