Module lpc55s6x_pac::sct0::event::ctrl

SCT event control register 0

Structs

IOSELR
Value of the field

MATCHMEMR
Value of the field

MATCHSELR
Value of the field

R
Value read from the register

STATEVR
Value of the field

W
Value to write to the register

_COMBMODEW
Proxy

_DIRECTIONW
Proxy

_HEVENTW
Proxy

_IOCONDW
Proxy

_IOSELW
Proxy

_MATCHMEMW
Proxy

_MATCHSELW
Proxy

_OUTSELW
Proxy

_STATELDW
Proxy

_STATEVW
Proxy

Enums

COMBMODER
Possible values of the field COMBMODE

COMBMODEW
Values that can be written to the field COMBMODE

DIRECTIONR
Possible values of the field DIRECTION

DIRECTIONW
Values that can be written to the field DIRECTION

HEVENTR
Possible values of the field HEVENT

HEVENTW
Values that can be written to the field HEVENT

IOCONDR
Possible values of the field IOCOND

IOCONDW
Values that can be written to the field IOCOND

OUTSELR
Possible values of the field OUTSEL

OUTSELW
Values that can be written to the field OUTSEL

STATELDR
Possible values of the field STATELD

STATELDW
Values that can be written to the field STATELD