Module lpc55s6x_pac::sct0::res

SCT conflict resolution register

Structs

R
Value read from the register

W
Value to write to the register

_O0RESW
Proxy

_O1RESW
Proxy

_O2RESW
Proxy

_O3RESW
Proxy

_O4RESW
Proxy

_O5RESW
Proxy

_O6RESW
Proxy

_O7RESW
Proxy

_O8RESW
Proxy

_O9RESW
Proxy

_O10RESW
Proxy

_O11RESW
Proxy

_O12RESW
Proxy

_O13RESW
Proxy

_O14RESW
Proxy

_O15RESW
Proxy

Enums

O0RESR
Possible values of the field O0RES

O0RESW
Values that can be written to the field O0RES

O1RESR
Possible values of the field O1RES

O1RESW
Values that can be written to the field O1RES

O2RESR
Possible values of the field O2RES

O2RESW
Values that can be written to the field O2RES

O3RESR
Possible values of the field O3RES

O3RESW
Values that can be written to the field O3RES

O4RESR
Possible values of the field O4RES

O4RESW
Values that can be written to the field O4RES

O5RESR
Possible values of the field O5RES

O5RESW
Values that can be written to the field O5RES

O6RESR
Possible values of the field O6RES

O6RESW
Values that can be written to the field O6RES

O7RESR
Possible values of the field O7RES

O7RESW
Values that can be written to the field O7RES

O8RESR
Possible values of the field O8RES

O8RESW
Values that can be written to the field O8RES

O9RESR
Possible values of the field O9RES

O9RESW
Values that can be written to the field O9RES

O10RESR
Possible values of the field O10RES

O10RESW
Values that can be written to the field O10RES

O11RESR
Possible values of the field O11RES

O11RESW
Values that can be written to the field O11RES

O12RESR
Possible values of the field O12RES

O12RESW
Values that can be written to the field O12RES

O13RESR
Possible values of the field O13RES

O13RESW
Values that can be written to the field O13RES

O14RESR
Possible values of the field O14RES

O14RESW
Values that can be written to the field O14RES

O15RESR
Possible values of the field O15RES

O15RESW
Values that can be written to the field O15RES