Module lpc55s6x_pac::syscon::ahbclkctrl0

AHB Clock control 0

Structs

R
Value read from the register

W
Value to write to the register

_ADCW
Proxy

_CRCGENW
Proxy

_DMA0W
Proxy

_FLASHW
Proxy

_FMCW
Proxy

_GINTW
Proxy

_GPIO0W
Proxy

_GPIO1W
Proxy

_GPIO2W
Proxy

_GPIO3W
Proxy

_IOCONW
Proxy

_MAILBOXW
Proxy

_MUX0W
Proxy

_PINTW
Proxy

_ROMW
Proxy

_RTCW
Proxy

_SRAM_CTRL1W
Proxy

_SRAM_CTRL2W
Proxy

_SRAM_CTRL3W
Proxy

_SRAM_CTRL4W
Proxy

_WWDTW
Proxy

Enums

ADCR
Possible values of the field ADC

ADCW
Values that can be written to the field ADC

CRCGENR
Possible values of the field CRCGEN

CRCGENW
Values that can be written to the field CRCGEN

DMA0R
Possible values of the field DMA0

DMA0W
Values that can be written to the field DMA0

FLASHR
Possible values of the field FLASH

FLASHW
Values that can be written to the field FLASH

FMCR
Possible values of the field FMC

FMCW
Values that can be written to the field FMC

GINTR
Possible values of the field GINT

GINTW
Values that can be written to the field GINT

GPIO0R
Possible values of the field GPIO0

GPIO0W
Values that can be written to the field GPIO0

GPIO1R
Possible values of the field GPIO1

GPIO1W
Values that can be written to the field GPIO1

GPIO2R
Possible values of the field GPIO2

GPIO2W
Values that can be written to the field GPIO2

GPIO3R
Possible values of the field GPIO3

GPIO3W
Values that can be written to the field GPIO3

IOCONR
Possible values of the field IOCON

IOCONW
Values that can be written to the field IOCON

MAILBOXR
Possible values of the field MAILBOX

MAILBOXW
Values that can be written to the field MAILBOX

MUX0R
Possible values of the field MUX0

MUX0W
Values that can be written to the field MUX0

PINTR
Possible values of the field PINT

PINTW
Values that can be written to the field PINT

ROMR
Possible values of the field ROM

ROMW
Values that can be written to the field ROM

RTCR
Possible values of the field RTC

RTCW
Values that can be written to the field RTC

SRAM_CTRL1R
Possible values of the field SRAM_CTRL1

SRAM_CTRL1W
Values that can be written to the field SRAM_CTRL1

SRAM_CTRL2R
Possible values of the field SRAM_CTRL2

SRAM_CTRL2W
Values that can be written to the field SRAM_CTRL2

SRAM_CTRL3R
Possible values of the field SRAM_CTRL3

SRAM_CTRL3W
Values that can be written to the field SRAM_CTRL3

SRAM_CTRL4R
Possible values of the field SRAM_CTRL4

SRAM_CTRL4W
Values that can be written to the field SRAM_CTRL4

WWDTR
Possible values of the field WWDT

WWDTW
Values that can be written to the field WWDT