Module lpc55s6x_pac::syscon::ahbclkctrl2

AHB Clock control 2

Structs

R
Value read from the register

W
Value to write to the register

_ANALOG_CTRLW
Proxy

_CAPT0W
Proxy

_CASPERW
Proxy

_COMPW
Proxy

_DMA1W
Proxy

_FREQMEW
Proxy

_GPIO4W
Proxy

_GPIO5W
Proxy

_GPIO_SECW
Proxy

_GPIO_SEC_INTW
Proxy

_HASH0W
Proxy

_HS_LSPIW
Proxy

_MUX1W
Proxy

_OTPW
Proxy

_PLULUTW
Proxy

_PQW
Proxy

_PUFW
Proxy

_RNGW
Proxy

_SDIOW
Proxy

_TIMER3W
Proxy

_TIMER4W
Proxy

_USB0_HOSTMW
Proxy

_USB0_HOSTSW
Proxy

_USB1_HOSTW
Proxy

_USB1_DEVW
Proxy

_USB1_RAMW
Proxy

_USB1_PHYW
Proxy

Enums

ANALOG_CTRLR
Possible values of the field ANALOG_CTRL

ANALOG_CTRLW
Values that can be written to the field ANALOG_CTRL

CAPT0R
Possible values of the field CAPT0

CAPT0W
Values that can be written to the field CAPT0

CASPERR
Possible values of the field CASPER

CASPERW
Values that can be written to the field CASPER

COMPR
Possible values of the field COMP

COMPW
Values that can be written to the field COMP

DMA1R
Possible values of the field DMA1

DMA1W
Values that can be written to the field DMA1

FREQMER
Possible values of the field FREQME

FREQMEW
Values that can be written to the field FREQME

GPIO4R
Possible values of the field GPIO4

GPIO4W
Values that can be written to the field GPIO4

GPIO5R
Possible values of the field GPIO5

GPIO5W
Values that can be written to the field GPIO5

GPIO_SECR
Possible values of the field GPIO_SEC

GPIO_SECW
Values that can be written to the field GPIO_SEC

GPIO_SEC_INTR
Possible values of the field GPIO_SEC_INT

GPIO_SEC_INTW
Values that can be written to the field GPIO_SEC_INT

HASH0R
Possible values of the field HASH0

HASH0W
Values that can be written to the field HASH0

HS_LSPIR
Possible values of the field HS_LSPI

HS_LSPIW
Values that can be written to the field HS_LSPI

MUX1R
Possible values of the field MUX1

MUX1W
Values that can be written to the field MUX1

OTPR
Possible values of the field OTP

OTPW
Values that can be written to the field OTP

PLULUTR
Possible values of the field PLULUT

PLULUTW
Values that can be written to the field PLULUT

PQR
Possible values of the field PQ

PQW
Values that can be written to the field PQ

PUFR
Possible values of the field PUF

PUFW
Values that can be written to the field PUF

RNGR
Possible values of the field RNG

RNGW
Values that can be written to the field RNG

SDIOR
Possible values of the field SDIO

SDIOW
Values that can be written to the field SDIO

TIMER3R
Possible values of the field TIMER3

TIMER3W
Values that can be written to the field TIMER3

TIMER4R
Possible values of the field TIMER4

TIMER4W
Values that can be written to the field TIMER4

USB0_HOSTMR
Possible values of the field USB0_HOSTM

USB0_HOSTSR
Possible values of the field USB0_HOSTS

USB0_HOSTMW
Values that can be written to the field USB0_HOSTM

USB0_HOSTSW
Values that can be written to the field USB0_HOSTS

USB1_HOSTR
Possible values of the field USB1_HOST

USB1_DEVR
Possible values of the field USB1_DEV

USB1_RAMR
Possible values of the field USB1_RAM

USB1_PHYR
Possible values of the field USB1_PHY

USB1_HOSTW
Values that can be written to the field USB1_HOST

USB1_DEVW
Values that can be written to the field USB1_DEV

USB1_RAMW
Values that can be written to the field USB1_RAM

USB1_PHYW
Values that can be written to the field USB1_PHY