Module lpc55s6x_pac::syscon::periphencfg

peripheral enable configuration -- FOR INTERNAL USE ONLY

Structs

R
Value read from the register

W
Value to write to the register

_ADCENW
Proxy

_HASHENW
Proxy

_PRINCEENW
Proxy

_PUFFENW
Proxy

_SCTENW
Proxy

_SDIOENW
Proxy

_USB0ENW
Proxy

_USB1ENW
Proxy

Enums

ADCENR
Possible values of the field ADCEN

ADCENW
Values that can be written to the field ADCEN

HASHENR
Possible values of the field HASHEN

HASHENW
Values that can be written to the field HASHEN

PRINCEENR
Possible values of the field PRINCEEN

PRINCEENW
Values that can be written to the field PRINCEEN

PUFFENR
Possible values of the field PUFFEN

PUFFENW
Values that can be written to the field PUFFEN

SCTENR
Possible values of the field SCTEN

SCTENW
Values that can be written to the field SCTEN

SDIOENR
Possible values of the field SDIOEN

SDIOENW
Values that can be written to the field SDIOEN

USB0ENR
Possible values of the field USB0EN

USB0ENW
Values that can be written to the field USB0EN

USB1ENR
Possible values of the field USB1EN

USB1ENW
Values that can be written to the field USB1EN