Module lpc55s6x_pac::syscon::pll1ctrl

PLL1 550m control

Structs

FRMCLKSTABLER
Value of the field

FRMENR
Value of the field

LIMUPOFFR
Value of the field

R
Value read from the register

SELIR
Value of the field

SELPR
Value of the field

SELRR
Value of the field

W
Value to write to the register

_BWDIRECTW
Proxy

_BYPASSPLLW
Proxy

_BYPASSPOSTDIV2W
Proxy

_BYPASSPOSTDIVW
Proxy

_BYPASSPREDIVW
Proxy

_CLKENW
Proxy

_FRMCLKSTABLEW
Proxy

_FRMENW
Proxy

_LIMUPOFFW
Proxy

_SELIW
Proxy

_SELPW
Proxy

_SELRW
Proxy

_SKEWENW
Proxy

Enums

BWDIRECTR
Possible values of the field BWDIRECT

BWDIRECTW
Values that can be written to the field BWDIRECT

BYPASSPLLR
Possible values of the field BYPASSPLL

BYPASSPLLW
Values that can be written to the field BYPASSPLL

BYPASSPOSTDIV2R
Possible values of the field BYPASSPOSTDIV2

BYPASSPOSTDIV2W
Values that can be written to the field BYPASSPOSTDIV2

BYPASSPOSTDIVR
Possible values of the field BYPASSPOSTDIV

BYPASSPOSTDIVW
Values that can be written to the field BYPASSPOSTDIV

BYPASSPREDIVR
Possible values of the field BYPASSPREDIV

BYPASSPREDIVW
Values that can be written to the field BYPASSPREDIV

CLKENR
Possible values of the field CLKEN

CLKENW
Values that can be written to the field CLKEN

SKEWENR
Possible values of the field SKEWEN

SKEWENW
Values that can be written to the field SKEWEN