Module lpc55s6x_pac::syscon::presetctrl0

Peripheral reset control 0

Structs

R
Value read from the register

W
Value to write to the register

_ADC_RSTW
Proxy

_CRCGEN_RSTW
Proxy

_DMA0_RSTW
Proxy

_FLASH_RSTW
Proxy

_FMC_RSTW
Proxy

_GINT_RSTW
Proxy

_GPIO0_RSTW
Proxy

_GPIO1_RSTW
Proxy

_GPIO2_RSTW
Proxy

_GPIO3_RSTW
Proxy

_IOCON_RSTW
Proxy

_MAILBOX_RSTW
Proxy

_MUX0_RSTW
Proxy

_PINT_RSTW
Proxy

_ROM_RSTW
Proxy

_RTC_RSTW
Proxy

_SRAM_CTRL1_RSTW
Proxy

_SRAM_CTRL2_RSTW
Proxy

_SRAM_CTRL3_RSTW
Proxy

_SRAM_CTRL4_RSTW
Proxy

_WWDT_RSTW
Proxy

Enums

ADC_RSTR
Possible values of the field ADC_RST

ADC_RSTW
Values that can be written to the field ADC_RST

CRCGEN_RSTR
Possible values of the field CRCGEN_RST

CRCGEN_RSTW
Values that can be written to the field CRCGEN_RST

DMA0_RSTR
Possible values of the field DMA0_RST

DMA0_RSTW
Values that can be written to the field DMA0_RST

FLASH_RSTR
Possible values of the field FLASH_RST

FLASH_RSTW
Values that can be written to the field FLASH_RST

FMC_RSTR
Possible values of the field FMC_RST

FMC_RSTW
Values that can be written to the field FMC_RST

GINT_RSTR
Possible values of the field GINT_RST

GINT_RSTW
Values that can be written to the field GINT_RST

GPIO0_RSTR
Possible values of the field GPIO0_RST

GPIO0_RSTW
Values that can be written to the field GPIO0_RST

GPIO1_RSTR
Possible values of the field GPIO1_RST

GPIO1_RSTW
Values that can be written to the field GPIO1_RST

GPIO2_RSTR
Possible values of the field GPIO2_RST

GPIO2_RSTW
Values that can be written to the field GPIO2_RST

GPIO3_RSTR
Possible values of the field GPIO3_RST

GPIO3_RSTW
Values that can be written to the field GPIO3_RST

IOCON_RSTR
Possible values of the field IOCON_RST

IOCON_RSTW
Values that can be written to the field IOCON_RST

MAILBOX_RSTR
Possible values of the field MAILBOX_RST

MAILBOX_RSTW
Values that can be written to the field MAILBOX_RST

MUX0_RSTR
Possible values of the field MUX0_RST

MUX0_RSTW
Values that can be written to the field MUX0_RST

PINT_RSTR
Possible values of the field PINT_RST

PINT_RSTW
Values that can be written to the field PINT_RST

ROM_RSTR
Possible values of the field ROM_RST

ROM_RSTW
Values that can be written to the field ROM_RST

RTC_RSTR
Possible values of the field RTC_RST

RTC_RSTW
Values that can be written to the field RTC_RST

SRAM_CTRL1_RSTR
Possible values of the field SRAM_CTRL1_RST

SRAM_CTRL1_RSTW
Values that can be written to the field SRAM_CTRL1_RST

SRAM_CTRL2_RSTR
Possible values of the field SRAM_CTRL2_RST

SRAM_CTRL2_RSTW
Values that can be written to the field SRAM_CTRL2_RST

SRAM_CTRL3_RSTR
Possible values of the field SRAM_CTRL3_RST

SRAM_CTRL3_RSTW
Values that can be written to the field SRAM_CTRL3_RST

SRAM_CTRL4_RSTR
Possible values of the field SRAM_CTRL4_RST

SRAM_CTRL4_RSTW
Values that can be written to the field SRAM_CTRL4_RST

WWDT_RSTR
Possible values of the field WWDT_RST

WWDT_RSTW
Values that can be written to the field WWDT_RST