Module lpc55s6x_pac::syscon::presetctrl2

Peripheral reset control 2

Structs

R
Value read from the register

W
Value to write to the register

_ANALOG_CTRL_RSTW
Proxy

_CAPT0_RSTW
Proxy

_CASPER_RSTW
Proxy

_COMP_RSTW
Proxy

_DMA1_RSTW
Proxy

_FREQME_RSTW
Proxy

_GPIO4_RSTW
Proxy

_GPIO5_RSTW
Proxy

_GPIO_SEC_INT_RSTW
Proxy

_GPIO_SEC_RSTW
Proxy

_HASH0_RSTW
Proxy

_HS_LSPI_RSTW
Proxy

_MUX1_RSTW
Proxy

_OTP_RSTW
Proxy

_PLULUT_RSTW
Proxy

_PQ_RSTW
Proxy

_PUF_RSTW
Proxy

_RNG_RSTW
Proxy

_SDIO_RSTW
Proxy

_TIMER3_RSTW
Proxy

_TIMER4_RSTW
Proxy

_USB0_HOSTM_RSTW
Proxy

_USB0_HOSTS_RSTW
Proxy

_USB1_HOST_RSTW
Proxy

_USB1_DEV_RSTW
Proxy

_USB1_RAM_RSTW
Proxy

_USB1_PHY_RSTW
Proxy

Enums

ANALOG_CTRL_RSTR
Possible values of the field ANALOG_CTRL_RST

ANALOG_CTRL_RSTW
Values that can be written to the field ANALOG_CTRL_RST

CAPT0_RSTR
Possible values of the field CAPT0_RST

CAPT0_RSTW
Values that can be written to the field CAPT0_RST

CASPER_RSTR
Possible values of the field CASPER_RST

CASPER_RSTW
Values that can be written to the field CASPER_RST

COMP_RSTR
Possible values of the field COMP_RST

COMP_RSTW
Values that can be written to the field COMP_RST

DMA1_RSTR
Possible values of the field DMA1_RST

DMA1_RSTW
Values that can be written to the field DMA1_RST

FREQME_RSTR
Possible values of the field FREQME_RST

FREQME_RSTW
Values that can be written to the field FREQME_RST

GPIO4_RSTR
Possible values of the field GPIO4_RST

GPIO4_RSTW
Values that can be written to the field GPIO4_RST

GPIO5_RSTR
Possible values of the field GPIO5_RST

GPIO5_RSTW
Values that can be written to the field GPIO5_RST

GPIO_SEC_INT_RSTR
Possible values of the field GPIO_SEC_INT_RST

GPIO_SEC_INT_RSTW
Values that can be written to the field GPIO_SEC_INT_RST

GPIO_SEC_RSTR
Possible values of the field GPIO_SEC_RST

GPIO_SEC_RSTW
Values that can be written to the field GPIO_SEC_RST

HASH0_RSTR
Possible values of the field HASH0_RST

HASH0_RSTW
Values that can be written to the field HASH0_RST

HS_LSPI_RSTR
Possible values of the field HS_LSPI_RST

HS_LSPI_RSTW
Values that can be written to the field HS_LSPI_RST

MUX1_RSTR
Possible values of the field MUX1_RST

MUX1_RSTW
Values that can be written to the field MUX1_RST

OTP_RSTR
Possible values of the field OTP_RST

OTP_RSTW
Values that can be written to the field OTP_RST

PLULUT_RSTR
Possible values of the field PLULUT_RST

PLULUT_RSTW
Values that can be written to the field PLULUT_RST

PQ_RSTR
Possible values of the field PQ_RST

PQ_RSTW
Values that can be written to the field PQ_RST

PUF_RSTR
Possible values of the field PUF_RST

PUF_RSTW
Values that can be written to the field PUF_RST

RNG_RSTR
Possible values of the field RNG_RST

RNG_RSTW
Values that can be written to the field RNG_RST

SDIO_RSTR
Possible values of the field SDIO_RST

SDIO_RSTW
Values that can be written to the field SDIO_RST

TIMER3_RSTR
Possible values of the field TIMER3_RST

TIMER3_RSTW
Values that can be written to the field TIMER3_RST

TIMER4_RSTR
Possible values of the field TIMER4_RST

TIMER4_RSTW
Values that can be written to the field TIMER4_RST

USB0_HOSTM_RSTR
Possible values of the field USB0_HOSTM_RST

USB0_HOSTS_RSTR
Possible values of the field USB0_HOSTS_RST

USB0_HOSTM_RSTW
Values that can be written to the field USB0_HOSTM_RST

USB0_HOSTS_RSTW
Values that can be written to the field USB0_HOSTS_RST

USB1_HOST_RSTR
Possible values of the field USB1_HOST_RST

USB1_DEV_RSTR
Possible values of the field USB1_DEV_RST

USB1_RAM_RSTR
Possible values of the field USB1_RAM_RST

USB1_PHY_RSTR
Possible values of the field USB1_PHY_RST

USB1_HOST_RSTW
Values that can be written to the field USB1_HOST_RST

USB1_DEV_RSTW
Values that can be written to the field USB1_DEV_RST

USB1_RAM_RSTW
Values that can be written to the field USB1_RAM_RST

USB1_PHY_RSTW
Values that can be written to the field USB1_PHY_RST