Module lpc55s6x_pac::syscon::starter1

Start logic wake-up enable register

Structs

R
Value read from the register

W
Value to write to the register

WAKEUPPADSR
Value of the field

_CASERW
Proxy

_CTIMER2W
Proxy

_CTIMER4W
Proxy

_GPIO_INT04W
Proxy

_GPIO_INT05W
Proxy

_GPIO_INT06W
Proxy

_GPIO_INT07W
Proxy

_LSPI_HSW
Proxy

_OS_EVENTW
Proxy

_PLUW
Proxy

_PQW
Proxy

_QDDKEYW
Proxy

_SDIOW
Proxy

_SDMA1W
Proxy

_SEC_GPIO_INT00W
Proxy

_SEC_GPIO_INT01W
Proxy

_SEC_HYPERVISOR_CALLW
Proxy

_SEC_VIOW
Proxy

_SHAW
Proxy

_USB1W
Proxy

_USB1_NEEDCLKW
Proxy

_WAKEUPPADSW
Proxy

Enums

CASERR
Possible values of the field CASER

CASERW
Values that can be written to the field CASER

CTIMER2R
Possible values of the field CTIMER2

CTIMER2W
Values that can be written to the field CTIMER2

CTIMER4R
Possible values of the field CTIMER4

CTIMER4W
Values that can be written to the field CTIMER4

GPIO_INT04R
Possible values of the field GPIO_INT04

GPIO_INT04W
Values that can be written to the field GPIO_INT04

GPIO_INT05R
Possible values of the field GPIO_INT05

GPIO_INT05W
Values that can be written to the field GPIO_INT05

GPIO_INT06R
Possible values of the field GPIO_INT06

GPIO_INT06W
Values that can be written to the field GPIO_INT06

GPIO_INT07R
Possible values of the field GPIO_INT07

GPIO_INT07W
Values that can be written to the field GPIO_INT07

LSPI_HSR
Possible values of the field LSPI_HS

LSPI_HSW
Values that can be written to the field LSPI_HS

OS_EVENTR
Possible values of the field OS_EVENT

OS_EVENTW
Values that can be written to the field OS_EVENT

PLUR
Possible values of the field PLU

PLUW
Values that can be written to the field PLU

PQR
Possible values of the field PQ

PQW
Values that can be written to the field PQ

QDDKEYR
Possible values of the field QDDKEY

QDDKEYW
Values that can be written to the field QDDKEY

SDIOR
Possible values of the field SDIO

SDIOW
Values that can be written to the field SDIO

SDMA1R
Possible values of the field SDMA1

SDMA1W
Values that can be written to the field SDMA1

SEC_GPIO_INT00R
Possible values of the field SEC_GPIO_INT00

SEC_GPIO_INT00W
Values that can be written to the field SEC_GPIO_INT00

SEC_GPIO_INT01R
Possible values of the field SEC_GPIO_INT01

SEC_GPIO_INT01W
Values that can be written to the field SEC_GPIO_INT01

SEC_HYPERVISOR_CALLR
Possible values of the field SEC_HYPERVISOR_CALL

SEC_HYPERVISOR_CALLW
Values that can be written to the field SEC_HYPERVISOR_CALL

SEC_VIOR
Possible values of the field SEC_VIO

SEC_VIOW
Values that can be written to the field SEC_VIO

SHAR
Possible values of the field SHA

SHAW
Values that can be written to the field SHA

USB1R
Possible values of the field USB1

USB1_NEEDCLKR
Possible values of the field USB1_NEEDCLK

USB1W
Values that can be written to the field USB1

USB1_NEEDCLKW
Values that can be written to the field USB1_NEEDCLK