Module lpc55s6x_pac::usart0::cfg

USART Configuration register. Basic USART configuration settings that typically are not changed during operation.

Structs

R
Value read from the register

W
Value to write to the register

_AUTOADDRW
Proxy

_CLKPOLW
Proxy

_CTSENW
Proxy

_DATALENW
Proxy

_ENABLEW
Proxy

_LINMODEW
Proxy

_LOOPW
Proxy

_MODE32KW
Proxy

_OEPOLW
Proxy

_OESELW
Proxy

_OETAW
Proxy

_PARITYSELW
Proxy

_RXPOLW
Proxy

_STOPLENW
Proxy

_SYNCENW
Proxy

_SYNCMSTW
Proxy

_TXPOLW
Proxy

Enums

AUTOADDRR
Possible values of the field AUTOADDR

AUTOADDRW
Values that can be written to the field AUTOADDR

CLKPOLR
Possible values of the field CLKPOL

CLKPOLW
Values that can be written to the field CLKPOL

CTSENR
Possible values of the field CTSEN

CTSENW
Values that can be written to the field CTSEN

DATALENR
Possible values of the field DATALEN

DATALENW
Values that can be written to the field DATALEN

ENABLER
Possible values of the field ENABLE

ENABLEW
Values that can be written to the field ENABLE

LINMODER
Possible values of the field LINMODE

LINMODEW
Values that can be written to the field LINMODE

LOOPR
Possible values of the field LOOP

LOOPW
Values that can be written to the field LOOP

MODE32KR
Possible values of the field MODE32K

MODE32KW
Values that can be written to the field MODE32K

OEPOLR
Possible values of the field OEPOL

OEPOLW
Values that can be written to the field OEPOL

OESELR
Possible values of the field OESEL

OESELW
Values that can be written to the field OESEL

OETAR
Possible values of the field OETA

OETAW
Values that can be written to the field OETA

PARITYSELR
Possible values of the field PARITYSEL

PARITYSELW
Values that can be written to the field PARITYSEL

RXPOLR
Possible values of the field RXPOL

RXPOLW
Values that can be written to the field RXPOL

STOPLENR
Possible values of the field STOPLEN

STOPLENW
Values that can be written to the field STOPLEN

SYNCENR
Possible values of the field SYNCEN

SYNCENW
Values that can be written to the field SYNCEN

SYNCMSTR
Possible values of the field SYNCMST

SYNCMSTW
Values that can be written to the field SYNCMST

TXPOLR
Possible values of the field TXPOL

TXPOLW
Values that can be written to the field TXPOL