Struct lpc55s6x_pac::usart0::cfg::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn enable(&mut self) -> _ENABLEW

Bit 0 - USART Enable.

pub fn datalen(&mut self) -> _DATALENW

Bits 2:3 - Selects the data size for the USART.

pub fn paritysel(&mut self) -> _PARITYSELW

Bits 4:5 - Selects what type of parity is used by the USART.

pub fn stoplen(&mut self) -> _STOPLENW

Bit 6 - Number of stop bits appended to transmitted data. Only a single stop bit is required for received data.

pub fn mode32k(&mut self) -> _MODE32KW

Bit 7 - Selects standard or 32 kHz clocking mode.

pub fn linmode(&mut self) -> _LINMODEW

Bit 8 - LIN break mode enable.

pub fn ctsen(&mut self) -> _CTSENW

Bit 9 - CTS Enable. Determines whether CTS is used for flow control. CTS can be from the input pin, or from the USART's own RTS if loopback mode is enabled.

pub fn syncen(&mut self) -> _SYNCENW

Bit 11 - Selects synchronous or asynchronous operation.

pub fn clkpol(&mut self) -> _CLKPOLW

Bit 12 - Selects the clock polarity and sampling edge of received data in synchronous mode.

pub fn syncmst(&mut self) -> _SYNCMSTW

Bit 14 - Synchronous mode Master select.

pub fn loop_(&mut self) -> _LOOPW

Bit 15 - Selects data loopback mode.

pub fn oeta(&mut self) -> _OETAW

Bit 18 - Output Enable Turnaround time enable for RS-485 operation.

pub fn autoaddr(&mut self) -> _AUTOADDRW

Bit 19 - Automatic Address matching enable.

pub fn oesel(&mut self) -> _OESELW

Bit 20 - Output Enable Select.

pub fn oepol(&mut self) -> _OEPOLW

Bit 21 - Output Enable Polarity.

pub fn rxpol(&mut self) -> _RXPOLW

Bit 22 - Receive data polarity.

pub fn txpol(&mut self) -> _TXPOLW

Bit 23 - Transmit data polarity.