Struct lpc55s6x_pac::usart0::stat::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn deltacts(&mut self) -> _DELTACTSW

Bit 5 - This bit is set when a change in the state is detected for the CTS flag above. This bit is cleared by software.

pub fn deltarxbrk(&mut self) -> _DELTARXBRKW

Bit 11 - This bit is set when a change in the state of receiver break detection occurs. Cleared by software.

pub fn start(&mut self) -> _STARTW

Bit 12 - This bit is set when a start is detected on the receiver input. Its purpose is primarily to allow wake-up from Deep-sleep or Power-down mode immediately when a start is detected. Cleared by software.

pub fn framerrint(&mut self) -> _FRAMERRINTW

Bit 13 - Framing Error interrupt flag. This flag is set when a character is received with a missing stop bit at the expected location. This could be an indication of a baud rate or configuration mismatch with the transmitting source.

pub fn parityerrint(&mut self) -> _PARITYERRINTW

Bit 14 - Parity Error interrupt flag. This flag is set when a parity error is detected in a received character.

pub fn rxnoiseint(&mut self) -> _RXNOISEINTW

Bit 15 - Received Noise interrupt flag. Three samples of received data are taken in order to determine the value of each received data bit, except in synchronous mode. This acts as a noise filter if one sample disagrees. This flag is set when a received data bit contains one disagreeing sample. This could indicate line noise, a baud rate or character format mismatch, or loss of synchronization during data reception.

pub fn aberr(&mut self) -> _ABERRW

Bit 16 - Auto baud Error. An auto baud error can occur if the BRG counts to its limit before the end of the start bit that is being measured, essentially an auto baud time-out.