Struct lpc55s6x_pac::usb0::devcmdstat::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn dev_addr(&mut self) -> _DEV_ADDRW

Bits 0:6 - USB device address. After bus reset, the address is reset to 0x00. If the enable bit is set, the device will respond on packets for function address DEV_ADDR. When receiving a SetAddress Control Request from the USB host, software must program the new address before completing the status phase of the SetAddress Control Request.

pub fn dev_en(&mut self) -> _DEV_ENW

Bit 7 - USB device enable. If this bit is set, the HW will start responding on packets for function address DEV_ADDR.

pub fn setup(&mut self) -> _SETUPW

Bit 8 - SETUP token received. If a SETUP token is received and acknowledged by the device, this bit is set. As long as this bit is set all received IN and OUT tokens will be NAKed by HW. SW must clear this bit by writing a one. If this bit is zero, HW will handle the tokens to the CTRL EP0 as indicated by the CTRL EP0 IN and OUT data information programmed by SW.

pub fn force_needclk(&mut self) -> _FORCE_NEEDCLKW

Bit 9 - Forces the NEEDCLK output to always be on:

pub fn lpm_sup(&mut self) -> _LPM_SUPW

Bit 11 - LPM Supported:

pub fn intonnak_ao(&mut self) -> _INTONNAK_AOW

Bit 12 - Interrupt on NAK for interrupt and bulk OUT EP

pub fn intonnak_ai(&mut self) -> _INTONNAK_AIW

Bit 13 - Interrupt on NAK for interrupt and bulk IN EP

pub fn intonnak_co(&mut self) -> _INTONNAK_COW

Bit 14 - Interrupt on NAK for control OUT EP

pub fn intonnak_ci(&mut self) -> _INTONNAK_CIW

Bit 15 - Interrupt on NAK for control IN EP

pub fn dcon(&mut self) -> _DCONW

Bit 16 - Device status - connect. The connect bit must be set by SW to indicate that the device must signal a connect. The pull-up resistor on USB_DP will be enabled when this bit is set and the VBUSDEBOUNCED bit is one.

pub fn dsus(&mut self) -> _DSUSW

Bit 17 - Device status - suspend. The suspend bit indicates the current suspend state. It is set to 1 when the device hasn't seen any activity on its upstream port for more than 3 milliseconds. It is reset to 0 on any activity. When the device is suspended (Suspend bit DSUS = 1) and the software writes a 0 to it, the device will generate a remote wake-up. This will only happen when the device is connected (Connect bit = 1). When the device is not connected or not suspended, a writing a 0 has no effect. Writing a 1 never has an effect.

pub fn lpm_sus(&mut self) -> _LPM_SUSW

Bit 19 - Device status - LPM Suspend. This bit represents the current LPM suspend state. It is set to 1 by HW when the device has acknowledged the LPM request from the USB host and the Token Retry Time of 10 ms has elapsed. When the device is in the LPM suspended state (LPM suspend bit = 1) and the software writes a zero to this bit, the device will generate a remote walk-up. Software can only write a zero to this bit when the LPM_REWP bit is set to 1. HW resets this bit when it receives a host initiated resume. HW only updates the LPM_SUS bit when the LPM_SUPP bit is equal to one.

pub fn dcon_c(&mut self) -> _DCON_CW

Bit 24 - Device status - connect change. The Connect Change bit is set when the device's pull-up resistor is disconnected because VBus disappeared. The bit is reset by writing a one to it.

pub fn dsus_c(&mut self) -> _DSUS_CW

Bit 25 - Device status - suspend change. The suspend change bit is set to 1 when the suspend bit toggles. The suspend bit can toggle because: - The device goes in the suspended state - The device is disconnected - The device receives resume signaling on its upstream port. The bit is reset by writing a one to it.

pub fn dres_c(&mut self) -> _DRES_CW

Bit 26 - Device status - reset change. This bit is set when the device received a bus reset. On a bus reset the device will automatically go to the default state (unconfigured and responding to address 0). The bit is reset by writing a one to it.