Struct lpc55s6x_pac::usbhsh::portsc1::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn ccs(&self) -> CCSR

Bit 0 - Current Connect Status: Logic 1 indicates a device is present on the port.

pub fn csc(&self) -> CSCR

Bit 1 - Connect Status Change: Logic 1 means that the value of CCS has changed.

pub fn ped(&self) -> PEDR

Bit 2 - Port Enabled/Disabled.

pub fn pedc(&self) -> PEDCR

Bit 3 - Port Enabled/Disabled Change: Logic 1 means that the value of PED has changed.

pub fn oca(&self) -> OCAR

Bit 4 - Over-current active: Logic 1 means that this port has an over-current condition.

pub fn occ(&self) -> OCCR

Bit 5 - Over-current change: Logic 1 means that the value of OCA has changed.

pub fn fpr(&self) -> FPRR

Bit 6 - Force Port Resume: Logic 1 means resume (K-state) detected or driven on the port.

pub fn susp(&self) -> SUSPR

Bit 7 - Suspend: Logic 1 means port is in the suspend state.

pub fn pr(&self) -> PRR

Bit 8 - Port Reset: Logic 1 means the port is in the reset state.

pub fn sus_l1(&self) -> SUS_L1R

Bit 9 - Suspend using L1 0b = Suspend using L2 1b = Suspend using L1 When this bit is set to a 1 and a non-zero value is specified in the Device Address field, the host controller will generate an LPM Token to enter the L1 state whenever software writes a one to the Suspend bit, as well as L1 exit timing during any device or host-initiated resume.

pub fn ls(&self) -> LSR

Bits 10:11 - Line Status: This field reflects the current logical levels of the DP (bit 11) and DM (bit 10) signal lines.

pub fn pp(&self) -> PPR

Bit 12 - Port Power: The function of this bit depends on the value of the Port Power Control (PPC) bit in the HCSPARAMS register.

pub fn pic(&self) -> PICR

Bits 14:15 - Port Indicator Control : Writing to this field has no effect if the P_INDICATOR bit in the HCSPARAMS register is logic 0.

pub fn ptc(&self) -> PTCR

Bits 16:19 - Port Test Control: A non-zero value indicates that the port is operating in the test mode as indicated by the value.

pub fn pspd(&self) -> PSPDR

Bits 20:21 - Port Speed: 00b: Low-speed 01b: Full-speed 10b: High-speed 11b: Reserved.

pub fn woo(&self) -> WOOR

Bit 22 - Wake on overcurrent enable: Writing this bit to a one enables the port to be sensitive to overcurrent conditions as wake-up events.

pub fn sus_stat(&self) -> SUS_STATR

Bits 23:24 - These two bits are used by software to determine whether the most recent L1 suspend request was successful: 00b: Success-state transition was successful (ACK) 01b: Not Yet - Device was unable to enter the L1 state at this time (NYET) 10b: Not supported - Device does not support the L1 state (STALL) 11b: Timeout/Error - Device failed to respond or an error occurred.

pub fn dev_add(&self) -> DEV_ADDR

Bits 25:31 - Device Address for LPM tokens.