Struct lpc55s6x_pac::usbhsh::usbcmd::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn rs(&self) -> RSR

Bit 0 - Run/Stop: 1b = Run.

pub fn hcreset(&self) -> HCRESETR

Bit 1 - Host Controller Reset: This control bit is used by the software to reset the host controller.

pub fn fls(&self) -> FLSR

Bits 2:3 - Frame List Size: This field specifies the size of the frame list.

pub fn lhcr(&self) -> LHCRR

Bit 7 - Light Host Controller Reset: This bit allows the driver software to reset the host controller without affecting the state of the ports.

pub fn atl_en(&self) -> ATL_ENR

Bit 8 - ATL List enabled.

pub fn iso_en(&self) -> ISO_ENR

Bit 9 - ISO List enabled.

pub fn int_en(&self) -> INT_ENR

Bit 10 - INT List enabled.

pub fn hird(&self) -> HIRDR

Bits 24:27 - Host-Initiated Resume Duration.

pub fn lpm_rwu(&self) -> LPM_RWUR

Bit 28 - bRemoteWake field.