Struct lpc55s6x_pac::usbphy::ctrl_tog::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn enhostdiscondetect(&self) -> ENHOSTDISCONDETECTR

Bit 1 - For host mode, enables high-speed disconnect detector

pub fn hostdiscondetect_irq(&self) -> HOSTDISCONDETECT_IRQR

Bit 3 - Indicates that the device has disconnected in High-Speed mode

pub fn endevplugindet(&self) -> ENDEVPLUGINDETR

Bit 4 - Enables non-standard resistive plugged-in detection This bit field controls connection of nominal 200kohm resistors to both the USB_DP and USB_DM pins as one method of detecting when a USB cable is attached in device mode

pub fn devplugin_irq(&self) -> DEVPLUGIN_IRQR

Bit 12 - Indicates that the device is connected

pub fn enutmilevel2(&self) -> ENUTMILEVEL2R

Bit 14 - Enables UTMI+ Level 2 operation for the USB HS PHY

pub fn enutmilevel3(&self) -> ENUTMILEVEL3R

Bit 15 - Enables UTMI+ Level 3 operation for the USB HS PHY

pub fn autoresume_en(&self) -> AUTORESUME_ENR

Bit 18 - Enable the auto resume feature, when set, HW will use 32KHz clock to send Resume to respond to the device remote wakeup(for host mode only)

pub fn enautoclr_clkgate(&self) -> ENAUTOCLR_CLKGATER

Bit 19 - Enables the feature to auto-clear the CLKGATE bit if there is wakeup event while USB is suspended

pub fn enautoclr_phy_pwd(&self) -> ENAUTOCLR_PHY_PWDR

Bit 20 - Enables the feature to auto-clear the PWD register bits in PWD if there is wakeup event while USB is suspended

pub fn fsdll_rst_en(&self) -> FSDLL_RST_ENR

Bit 24 - Enables the feature to reset the FSDLL lock detection logic at the end of each TX packet.

pub fn otg_id_value(&self) -> OTG_ID_VALUER

Bit 27 - Indicates the results of USB_ID pin while monitoring the cable plugged into the Micro- or Mini-AB receptacle

pub fn host_force_ls_se0(&self) -> HOST_FORCE_LS_SE0R

Bit 28 - Forces the next FS packet that is transmitted to have a EOP with low-speed timing

pub fn utmi_suspendm(&self) -> UTMI_SUSPENDMR

Bit 29 - Used by the PHY to indicate a powered-down state

pub fn clkgate(&self) -> CLKGATER

Bit 30 - Gate UTMI Clocks

pub fn sftrst(&self) -> SFTRSTR

Bit 31 - Writing a 1 to this bit will soft-reset the PWD, TX, RX, and CTRL registers