Module lpc55s6x_pac::usbphy::pll_sic

USB PHY PLL Control/Status Register

Structs

MISC2_CONTROL0R
Value of the field

PLL_BYPASSR
Value of the field

PLL_ENABLER
Value of the field

PLL_EN_USB_CLKSR
Value of the field

PLL_POWERR
Value of the field

PLL_REG_ENABLER
Value of the field

R
Value read from the register

REFBIAS_PWDR
Value of the field

W
Value to write to the register

_MISC2_CONTROL0W
Proxy

_PLL_BYPASSW
Proxy

_PLL_DIV_SELW
Proxy

_PLL_ENABLEW
Proxy

_PLL_EN_USB_CLKSW
Proxy

_PLL_LOCKW
Proxy

_PLL_POWERW
Proxy

_PLL_REG_ENABLEW
Proxy

_REFBIAS_PWDW
Proxy

_REFBIAS_PWD_SELW
Proxy

Enums

PLL_DIV_SELR
Possible values of the field PLL_DIV_SEL

PLL_DIV_SELW
Values that can be written to the field PLL_DIV_SEL

PLL_LOCKR
Possible values of the field PLL_LOCK

PLL_LOCKW
Values that can be written to the field PLL_LOCK

REFBIAS_PWD_SELR
Possible values of the field REFBIAS_PWD_SEL

REFBIAS_PWD_SELW
Values that can be written to the field REFBIAS_PWD_SEL