Module lpc55s6x_pac::usbphy::usb1_vbus_detect

USB PHY VBUS Detect Control Register

Structs

AVALID_OVERRIDER
Value of the field

BVALID_OVERRIDER
Value of the field

R
Value read from the register

SESSEND_OVERRIDER
Value of the field

VBUSVALID_OVERRIDER
Value of the field

W
Value to write to the register

_AVALID_OVERRIDEW
Proxy

_BVALID_OVERRIDEW
Proxy

_DISCHARGE_VBUSW
Proxy

_EN_CHARGER_RESISTORW
Proxy

_PWRUP_CMPSW
Proxy

_SESSEND_OVERRIDEW
Proxy

_VBUSVALID_OVERRIDEW
Proxy

_VBUSVALID_SELW
Proxy

_VBUSVALID_THRESHW
Proxy

_VBUSVALID_TO_SESSVALIDW
Proxy

_VBUS_OVERRIDE_ENW
Proxy

_VBUS_SOURCE_SELW
Proxy

Enums

DISCHARGE_VBUSR
Possible values of the field DISCHARGE_VBUS

DISCHARGE_VBUSW
Values that can be written to the field DISCHARGE_VBUS

EN_CHARGER_RESISTORR
Possible values of the field EN_CHARGER_RESISTOR

EN_CHARGER_RESISTORW
Values that can be written to the field EN_CHARGER_RESISTOR

PWRUP_CMPSR
Possible values of the field PWRUP_CMPS

PWRUP_CMPSW
Values that can be written to the field PWRUP_CMPS

VBUSVALID_SELR
Possible values of the field VBUSVALID_SEL

VBUSVALID_SELW
Values that can be written to the field VBUSVALID_SEL

VBUSVALID_THRESHR
Possible values of the field VBUSVALID_THRESH

VBUSVALID_THRESHW
Values that can be written to the field VBUSVALID_THRESH

VBUSVALID_TO_SESSVALIDR
Possible values of the field VBUSVALID_TO_SESSVALID

VBUSVALID_TO_SESSVALIDW
Values that can be written to the field VBUSVALID_TO_SESSVALID

VBUS_OVERRIDE_ENR
Possible values of the field VBUS_OVERRIDE_EN

VBUS_OVERRIDE_ENW
Values that can be written to the field VBUS_OVERRIDE_EN

VBUS_SOURCE_SELR
Possible values of the field VBUS_SOURCE_SEL

VBUS_SOURCE_SELW
Values that can be written to the field VBUS_SOURCE_SEL