Struct lpc55s6x_pac::usbphy::usb1_vbus_detect_clr::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn vbusvalid_thresh(&mut self) -> _VBUSVALID_THRESHW

Bits 0:2 - Sets the threshold for the VBUSVALID comparator

pub fn vbus_override_en(&mut self) -> _VBUS_OVERRIDE_ENW

Bit 3 - VBUS detect signal override enable

pub fn sessend_override(&mut self) -> _SESSEND_OVERRIDEW

Bit 4 - Override value for SESSEND The bit field provides the value for USB1_VBUS_DET_STAT[0] if USB_VBUS_DETECT[3] is set to value 1'b1

pub fn bvalid_override(&mut self) -> _BVALID_OVERRIDEW

Bit 5 - Override value for B-Device Session Valid The bit field provides the value for USB1_VBUS_DET_STAT[1] if USB_VBUS_DETECT[3] is set to value 1'b1

pub fn avalid_override(&mut self) -> _AVALID_OVERRIDEW

Bit 6 - Override value for A-Device Session Valid The bit field provides the value for USB1_VBUS_DET_STAT[2] if USB_VBUS_DETECT[3] is set to value 1'b1

pub fn vbusvalid_override(&mut self) -> _VBUSVALID_OVERRIDEW

Bit 7 - Override value for VBUS_VALID signal sent to USB controller The bit field provides the value for VBUS_VALID reported to the USB controller if the value of USB1_VBUS_DETECT[3] is set to 1'b1

pub fn vbusvalid_sel(&mut self) -> _VBUSVALID_SELW

Bit 8 - Selects the source of the VBUS_VALID signal reported to the USB controller This is one of the bit fields that selects the source of the VBUS_VALID signal reported to the USB controller

pub fn vbus_source_sel(&mut self) -> _VBUS_SOURCE_SELW

Bits 9:10 - Selects the source of the VBUS_VALID signal reported to the USB controller This is one of the bit fields that selects the source of the VBUS_VALID signal reported to the USB controller

pub fn vbusvalid_to_sessvalid(&mut self) -> _VBUSVALID_TO_SESSVALIDW

Bit 18 - Selects the comparator used for VBUS_VALID This bit field controls the comparator used to report the VBUS_VALID results in USB1_VBUS_DETECT[3] between the VBUS_VALID comparator and the Session Valid comparator

pub fn pwrup_cmps(&mut self) -> _PWRUP_CMPSW

Bit 20 - Enables the VBUS_VALID comparator Powers up the comparator used for the VBUS_VALID detector

pub fn discharge_vbus(&mut self) -> _DISCHARGE_VBUSW

Bit 26 - Controls VBUS discharge resistor This bit field controls a nominal 22kohm resistor between the USB1_VBUS pin and ground

pub fn en_charger_resistor(&mut self) -> _EN_CHARGER_RESISTORW

Bit 31 - Enables resistors used for an older method of resistive battery charger detection