Struct lpc55s6x_pac::usbphy::usb1_vbus_detect_set::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn vbusvalid_thresh(&self) -> VBUSVALID_THRESHR

Bits 0:2 - Sets the threshold for the VBUSVALID comparator

pub fn vbus_override_en(&self) -> VBUS_OVERRIDE_ENR

Bit 3 - VBUS detect signal override enable

pub fn sessend_override(&self) -> SESSEND_OVERRIDER

Bit 4 - Override value for SESSEND The bit field provides the value for USB1_VBUS_DET_STAT[0] if USB_VBUS_DETECT[3] is set to value 1'b1

pub fn bvalid_override(&self) -> BVALID_OVERRIDER

Bit 5 - Override value for B-Device Session Valid The bit field provides the value for USB1_VBUS_DET_STAT[1] if USB_VBUS_DETECT[3] is set to value 1'b1

pub fn avalid_override(&self) -> AVALID_OVERRIDER

Bit 6 - Override value for A-Device Session Valid The bit field provides the value for USB1_VBUS_DET_STAT[2] if USB_VBUS_DETECT[3] is set to value 1'b1

pub fn vbusvalid_override(&self) -> VBUSVALID_OVERRIDER

Bit 7 - Override value for VBUS_VALID signal sent to USB controller The bit field provides the value for VBUS_VALID reported to the USB controller if the value of USB1_VBUS_DETECT[3] is set to 1'b1

pub fn vbusvalid_sel(&self) -> VBUSVALID_SELR

Bit 8 - Selects the source of the VBUS_VALID signal reported to the USB controller This is one of the bit fields that selects the source of the VBUS_VALID signal reported to the USB controller

pub fn vbus_source_sel(&self) -> VBUS_SOURCE_SELR

Bits 9:10 - Selects the source of the VBUS_VALID signal reported to the USB controller This is one of the bit fields that selects the source of the VBUS_VALID signal reported to the USB controller

pub fn vbusvalid_to_sessvalid(&self) -> VBUSVALID_TO_SESSVALIDR

Bit 18 - Selects the comparator used for VBUS_VALID This bit field controls the comparator used to report the VBUS_VALID results in USB1_VBUS_DETECT[3] between the VBUS_VALID comparator and the Session Valid comparator

pub fn pwrup_cmps(&self) -> PWRUP_CMPSR

Bit 20 - Enables the VBUS_VALID comparator Powers up the comparator used for the VBUS_VALID detector

pub fn discharge_vbus(&self) -> DISCHARGE_VBUSR

Bit 26 - Controls VBUS discharge resistor This bit field controls a nominal 22kohm resistor between the USB1_VBUS pin and ground

pub fn en_charger_resistor(&self) -> EN_CHARGER_RESISTORR

Bit 31 - Enables resistors used for an older method of resistive battery charger detection